Refactor trainer-config.ts
