+++ src/data/trainer-config.ts
@@ -1,22 +1,25 @@
 import BattleScene, { startingWave } from "#app/battle-scene";
 import { ModifierTypeFunc, modifierTypes } from "#app/modifier/modifier-type";
-import { EnemyPokemon, PokemonMove } from "#app/field/pokemon";
+import { EnemyPokemon } from "#app/field/pokemon";
 import * as Utils from "#app/utils";
-import { PokeballType } from "#enums/pokeball";
-import { pokemonEvolutions, pokemonPrevolutions } from "#app/data/balance/pokemon-evolutions";
 import PokemonSpecies, { getPokemonSpecies, PokemonSpeciesFilter } from "#app/data/pokemon-species";
-import { tmSpecies } from "#app/data/balance/tms";
 import { Type } from "#enums/type";
 import { doubleBattleDialogue } from "#app/data/dialogue";
 import { PersistentModifier } from "#app/modifier/modifier";
 import { TrainerVariant } from "#app/field/trainer";
 import { getIsInitialized, initI18n } from "#app/plugins/i18n";
 import i18next from "i18next";
-import { Moves } from "#enums/moves";
 import { PartyMemberStrength } from "#enums/party-member-strength";
 import { Species } from "#enums/species";
 import { TrainerType } from "#enums/trainer-type";
-import { Gender } from "#app/data/gender";
+import { genericTrainerConfigs } from "./balance/trainers/generic-trainer-config";
+import { evilTeamTrainerConfigs } from "./balance/trainers/evil-team-trainer-config";
+import { gymLeaderTrainerConfigs } from "./balance/trainers/gym-leader-trainer-config";
+import { eliteFourTrainerConfigs } from "./balance/trainers/elite-four-trainer-config";
+import { chmapionTrainerConfigs } from "./balance/trainers/champion-trainer-config";
+import { rivalTrainerConfigs } from "./balance/trainers/rival-trainer-config";
+import { evilBossTrainerConfigs } from "./balance/trainers/evil-boss-trainer-config";
+import { mysteryEncounterTrainerConfigs } from "./balance/trainers/mystery-encounter-trainer-config";
 
 /** Minimum BST for Pokemon generated onto the Elite Four's teams */
 const ELITE_FOUR_MINIMUM_BST = 460;
@@ -41,38 +44,48 @@ export enum TrainerSlot {
   TRAINER_PARTNER,
 }
 
+/**
+ * Class that represents a trainer's party
+ * @param size - the number of Pokemon in the party (1-6)
+ * @param strength - how strong the Pokemon are
+ * @param sameSpecies - whether or not all Pokemon in this party ar ethe same
+ * @param balanced - whether or not to add an extra retry check in trainer.ts if there are any duplicate types
+ */
 export class TrainerPartyTemplate {
-  public size: integer;
+  public size: number;
   public strength: PartyMemberStrength;
   public sameSpecies: boolean;
   public balanced: boolean;
 
-  constructor(size: integer, strength: PartyMemberStrength, sameSpecies?: boolean, balanced?: boolean) {
+  constructor(size: number, strength: PartyMemberStrength, sameSpecies?: boolean, balanced?: boolean) {
     this.size = size;
     this.strength = strength;
     this.sameSpecies = !!sameSpecies;
     this.balanced = !!balanced;
   }
 
-  getStrength(index: integer): PartyMemberStrength {
+  getStrength(index: number): PartyMemberStrength {
     return this.strength;
   }
 
-  isSameSpecies(index: integer): boolean {
+  isSameSpecies(index: number): boolean {
     return this.sameSpecies;
   }
 
-  isBalanced(index: integer): boolean {
+  isBalanced(index: number): boolean {
     return this.balanced;
   }
 }
 
+/**
+ * A TrainerPartyTemplate that holds other TrainerPartyTemplates
+ */
 export class TrainerPartyCompoundTemplate extends TrainerPartyTemplate {
   public templates: TrainerPartyTemplate[];
 
   constructor(...templates: TrainerPartyTemplate[]) {
     super(
-      templates.reduce((total: integer, template: TrainerPartyTemplate) => {
+      templates.reduce((total: number, template: TrainerPartyTemplate) => {
         total += template.size;
         return total;
       }, 0),
@@ -81,7 +94,7 @@ export class TrainerPartyCompoundTemplate extends TrainerPartyTemplate {
     this.templates = templates;
   }
 
-  getStrength(index: integer): PartyMemberStrength {
+  getStrength(index: number): PartyMemberStrength {
     let t = 0;
     for (const template of this.templates) {
       if (t + template.size > index) {
@@ -93,7 +106,7 @@ export class TrainerPartyCompoundTemplate extends TrainerPartyTemplate {
     return super.getStrength(index);
   }
 
-  isSameSpecies(index: integer): boolean {
+  isSameSpecies(index: number): boolean {
     let t = 0;
     for (const template of this.templates) {
       if (t + template.size > index) {
@@ -105,7 +118,7 @@ export class TrainerPartyCompoundTemplate extends TrainerPartyTemplate {
     return super.isSameSpecies(index);
   }
 
-  isBalanced(index: integer): boolean {
+  isBalanced(index: number): boolean {
     let t = 0;
     for (const template of this.templates) {
       if (t + template.size > index) {
@@ -256,11 +269,11 @@ export const trainerPartyTemplates = {
 };
 
 type PartyTemplateFunc = (scene: BattleScene) => TrainerPartyTemplate;
-type PartyMemberFunc = (scene: BattleScene, level: integer, strength: PartyMemberStrength) => EnemyPokemon;
+type PartyMemberFunc = (scene: BattleScene, level: number, strength: PartyMemberStrength) => EnemyPokemon;
 type GenModifiersFunc = (party: EnemyPokemon[]) => PersistentModifier[];
 
 export interface PartyMemberFuncs {
-  [key: integer]: PartyMemberFunc;
+  [key: number]: PartyMemberFunc;
 }
 
 export class TrainerConfig {
@@ -1521,33 +1534,11 @@ export class TrainerConfig {
   }
 }
 
-let t = 0;
-
 interface TrainerConfigs {
-  [key: integer]: TrainerConfig;
+  [key: number]: TrainerConfig;
 }
 
-/**
- * The function to get variable strength grunts
- * @param scene the singleton scene being passed in
- * @returns the correct TrainerPartyTemplate
- */
-function getEvilGruntPartyTemplate(scene: BattleScene): TrainerPartyTemplate {
-  const waveIndex = scene.currentBattle?.waveIndex;
-  if (waveIndex < 40) {
-    return trainerPartyTemplates.TWO_AVG;
-  } else if (waveIndex < 63) {
-    return trainerPartyTemplates.THREE_AVG;
-  } else if (waveIndex < 65) {
-    return trainerPartyTemplates.TWO_AVG_ONE_STRONG;
-  } else if (waveIndex < 112) {
-    return trainerPartyTemplates.GYM_LEADER_4; // 3avg 1 strong 1 stronger
-  } else {
-    return trainerPartyTemplates.GYM_LEADER_5; // 3 avg 2 strong 1 stronger
-  }
-}
-
-function getWavePartyTemplate(scene: BattleScene, ...templates: TrainerPartyTemplate[]) {
+export function getWavePartyTemplate(scene: BattleScene, ...templates: TrainerPartyTemplate[]) {
   return templates[
     Math.min(
       Math.max(
@@ -1608,7 +1599,7 @@ export function getRandomPartyMemberFunc(
   };
 }
 
-function getSpeciesFilterRandomPartyMemberFunc(
+export function getSpeciesFilterRandomPartyMemberFunc(
   originalSpeciesFilter: PokemonSpeciesFilter,
   trainerSlot: TrainerSlot = TrainerSlot.TRAINER,
   allowLegendaries?: boolean,
@@ -1648,3760 +1639,13 @@ function getRandomTeraModifiers(party: EnemyPokemon[], count: integer, types?: T
   return ret;
 }
 
-type SignatureSpecies = {
-  [key in string]: (Species | Species[])[];
-};
-
-/*
- * The signature species for each Gym Leader, Elite Four member, and Champion.
- * The key is the trainer type, and the value is an array of Species or Species arrays.
- * This is in a separate const so it can be accessed from other places and not just the trainerConfigs
- */
-export const signatureSpecies: SignatureSpecies = {
-  BROCK: [Species.GEODUDE, Species.ONIX],
-  MISTY: [Species.STARYU, Species.PSYDUCK],
-  LT_SURGE: [Species.VOLTORB, Species.PIKACHU, Species.ELECTABUZZ],
-  ERIKA: [Species.ODDISH, Species.BELLSPROUT, Species.TANGELA, Species.HOPPIP],
-  JANINE: [Species.VENONAT, Species.SPINARAK, Species.ZUBAT],
-  SABRINA: [Species.ABRA, Species.MR_MIME, Species.ESPEON],
-  BLAINE: [Species.GROWLITHE, Species.PONYTA, Species.MAGMAR],
-  GIOVANNI: [Species.SANDILE, Species.MURKROW, Species.NIDORAN_M, Species.NIDORAN_F],
-  FALKNER: [Species.PIDGEY, Species.HOOTHOOT, Species.DODUO],
-  BUGSY: [Species.SCYTHER, Species.HERACROSS, Species.SHUCKLE, Species.PINSIR],
-  WHITNEY: [Species.GIRAFARIG, Species.MILTANK],
-  MORTY: [Species.GASTLY, Species.MISDREAVUS, Species.SABLEYE],
-  CHUCK: [Species.POLIWRATH, Species.MANKEY],
-  JASMINE: [Species.MAGNEMITE, Species.STEELIX],
-  PRYCE: [Species.SEEL, Species.SWINUB],
-  CLAIR: [Species.DRATINI, Species.HORSEA, Species.GYARADOS],
-  ROXANNE: [Species.GEODUDE, Species.NOSEPASS],
-  BRAWLY: [Species.MACHOP, Species.MAKUHITA],
-  WATTSON: [Species.MAGNEMITE, Species.VOLTORB, Species.ELECTRIKE],
-  FLANNERY: [Species.SLUGMA, Species.TORKOAL, Species.NUMEL],
-  NORMAN: [Species.SLAKOTH, Species.SPINDA, Species.CHANSEY, Species.KANGASKHAN],
-  WINONA: [Species.SWABLU, Species.WINGULL, Species.TROPIUS, Species.SKARMORY],
-  TATE: [Species.SOLROCK, Species.NATU, Species.CHIMECHO, Species.GALLADE],
-  LIZA: [Species.LUNATONE, Species.SPOINK, Species.BALTOY, Species.GARDEVOIR],
-  JUAN: [Species.HORSEA, Species.BARBOACH, Species.SPHEAL, Species.RELICANTH],
-  ROARK: [Species.CRANIDOS, Species.LARVITAR, Species.GEODUDE],
-  GARDENIA: [Species.ROSELIA, Species.TANGELA, Species.TURTWIG],
-  MAYLENE: [Species.LUCARIO, Species.MEDITITE, Species.CHIMCHAR],
-  CRASHER_WAKE: [Species.BUIZEL, Species.MAGIKARP, Species.PIPLUP],
-  FANTINA: [Species.MISDREAVUS, Species.DRIFLOON, Species.SPIRITOMB],
-  BYRON: [Species.SHIELDON, Species.BRONZOR, Species.AGGRON],
-  CANDICE: [Species.SNEASEL, Species.SNOVER, Species.SNORUNT],
-  VOLKNER: [Species.SHINX, Species.CHINCHOU, Species.ROTOM],
-  CILAN: [Species.PANSAGE, Species.COTTONEE, Species.PETILIL],
-  CHILI: [Species.PANSEAR, Species.DARUMAKA, Species.HEATMOR],
-  CRESS: [Species.PANPOUR, Species.BASCULIN, Species.TYMPOLE],
-  CHEREN: [Species.LILLIPUP, Species.MINCCINO, Species.PATRAT],
-  LENORA: [Species.KANGASKHAN, Species.DEERLING, Species.AUDINO],
-  ROXIE: [Species.VENIPEDE, Species.TRUBBISH, Species.SKORUPI],
-  BURGH: [Species.SEWADDLE, Species.SHELMET, Species.KARRABLAST],
-  ELESA: [Species.EMOLGA, Species.BLITZLE, Species.JOLTIK],
-  CLAY: [Species.DRILBUR, Species.SANDILE, Species.GOLETT],
-  SKYLA: [Species.DUCKLETT, Species.WOOBAT, Species.RUFFLET],
-  BRYCEN: [Species.CRYOGONAL, Species.VANILLITE, Species.CUBCHOO],
-  DRAYDEN: [Species.DRUDDIGON, Species.AXEW, Species.DEINO],
-  MARLON: [Species.WAILMER, Species.FRILLISH, Species.TIRTOUGA],
-  VIOLA: [Species.SURSKIT, Species.SCATTERBUG],
-  GRANT: [Species.AMAURA, Species.TYRUNT],
-  KORRINA: [Species.HAWLUCHA, Species.LUCARIO, Species.MIENFOO],
-  RAMOS: [Species.SKIDDO, Species.HOPPIP, Species.BELLSPROUT],
-  CLEMONT: [Species.HELIOPTILE, Species.MAGNEMITE, Species.EMOLGA],
-  VALERIE: [Species.SYLVEON, Species.MAWILE, Species.MR_MIME],
-  OLYMPIA: [Species.ESPURR, Species.SIGILYPH, Species.SLOWKING],
-  WULFRIC: [Species.BERGMITE, Species.SNOVER, Species.CRYOGONAL],
-  MILO: [Species.GOSSIFLEUR, Species.APPLIN, Species.BOUNSWEET],
-  NESSA: [Species.CHEWTLE, Species.ARROKUDA, Species.WIMPOD],
-  KABU: [Species.SIZZLIPEDE, Species.VULPIX, Species.TORKOAL],
-  BEA: [Species.GALAR_FARFETCHD, Species.MACHOP, Species.CLOBBOPUS],
-  ALLISTER: [Species.GALAR_YAMASK, Species.GALAR_CORSOLA, Species.GASTLY],
-  OPAL: [Species.MILCERY, Species.TOGETIC, Species.GALAR_WEEZING],
-  BEDE: [Species.HATENNA, Species.GALAR_PONYTA, Species.GARDEVOIR],
-  GORDIE: [Species.ROLYCOLY, Species.STONJOURNER, Species.BINACLE],
-  MELONY: [Species.SNOM, Species.GALAR_DARUMAKA, Species.GALAR_MR_MIME],
-  PIERS: [Species.GALAR_ZIGZAGOON, Species.SCRAGGY, Species.INKAY],
-  MARNIE: [Species.IMPIDIMP, Species.PURRLOIN, Species.MORPEKO],
-  RAIHAN: [Species.DURALUDON, Species.TURTONATOR, Species.GOOMY],
-  KATY: [Species.NYMBLE, Species.TAROUNTULA, Species.HERACROSS],
-  BRASSIUS: [Species.SMOLIV, Species.SHROOMISH, Species.ODDISH],
-  IONO: [Species.TADBULB, Species.WATTREL, Species.VOLTORB],
-  KOFU: [Species.VELUZA, Species.WIGLETT, Species.WINGULL],
-  LARRY: [Species.STARLY, Species.DUNSPARCE, Species.KOMALA],
-  RYME: [Species.GREAVARD, Species.SHUPPET, Species.MIMIKYU],
-  TULIP: [Species.GIRAFARIG, Species.FLITTLE, Species.RALTS],
-  GRUSHA: [Species.CETODDLE, Species.ALOLA_VULPIX, Species.CUBCHOO],
-  LORELEI: [
-    Species.JYNX,
-    [Species.SLOWBRO, Species.GALAR_SLOWBRO],
-    Species.LAPRAS,
-    [Species.ALOLA_SANDSLASH, Species.CLOYSTER],
-  ],
-  BRUNO: [Species.MACHAMP, Species.HITMONCHAN, Species.HITMONLEE, [Species.ALOLA_GOLEM, Species.GOLEM]],
-  AGATHA: [Species.GENGAR, [Species.ARBOK, Species.WEEZING], Species.CROBAT, Species.ALOLA_MAROWAK],
-  LANCE: [Species.DRAGONITE, Species.GYARADOS, Species.AERODACTYL, Species.ALOLA_EXEGGUTOR],
-  WILL: [Species.XATU, Species.JYNX, [Species.SLOWBRO, Species.SLOWKING], Species.EXEGGUTOR],
-  KOGA: [[Species.WEEZING, Species.MUK], [Species.VENOMOTH, Species.ARIADOS], Species.CROBAT, Species.TENTACRUEL],
-  KAREN: [Species.UMBREON, Species.HONCHKROW, Species.HOUNDOOM, Species.WEAVILE],
-  SIDNEY: [
-    [Species.SHIFTRY, Species.CACTURNE],
-    [Species.SHARPEDO, Species.CRAWDAUNT],
-    Species.ABSOL,
-    Species.MIGHTYENA,
-  ],
-  PHOEBE: [Species.SABLEYE, Species.DUSKNOIR, Species.BANETTE, [Species.MISMAGIUS, Species.DRIFBLIM]],
-  GLACIA: [Species.GLALIE, Species.WALREIN, Species.FROSLASS, Species.ABOMASNOW],
-  DRAKE: [Species.ALTARIA, Species.SALAMENCE, Species.FLYGON, Species.KINGDRA],
-  AARON: [[Species.SCIZOR, Species.KLEAVOR], Species.HERACROSS, [Species.VESPIQUEN, Species.YANMEGA], Species.DRAPION],
-  BERTHA: [Species.WHISCASH, Species.HIPPOWDON, Species.GLISCOR, Species.RHYPERIOR],
-  FLINT: [
-    [Species.FLAREON, Species.RAPIDASH],
-    Species.MAGMORTAR,
-    [Species.STEELIX, Species.LOPUNNY],
-    Species.INFERNAPE,
-  ],
-  LUCIAN: [Species.MR_MIME, Species.GALLADE, Species.BRONZONG, [Species.ALAKAZAM, Species.ESPEON]],
-  SHAUNTAL: [Species.COFAGRIGUS, Species.CHANDELURE, Species.GOLURK, Species.JELLICENT],
-  MARSHAL: [Species.CONKELDURR, Species.MIENSHAO, Species.THROH, Species.SAWK],
-  GRIMSLEY: [Species.LIEPARD, Species.KINGAMBIT, Species.SCRAFTY, Species.KROOKODILE],
-  CAITLIN: [Species.MUSHARNA, Species.GOTHITELLE, Species.SIGILYPH, Species.REUNICLUS],
-  MALVA: [Species.PYROAR, Species.TORKOAL, Species.CHANDELURE, Species.TALONFLAME],
-  SIEBOLD: [Species.CLAWITZER, Species.GYARADOS, Species.BARBARACLE, Species.STARMIE],
-  WIKSTROM: [Species.KLEFKI, Species.PROBOPASS, Species.SCIZOR, Species.AEGISLASH],
-  DRASNA: [Species.DRAGALGE, Species.DRUDDIGON, Species.ALTARIA, Species.NOIVERN],
-  HALA: [Species.HARIYAMA, Species.BEWEAR, Species.CRABOMINABLE, [Species.POLIWRATH, Species.ANNIHILAPE]],
-  MOLAYNE: [Species.KLEFKI, Species.MAGNEZONE, Species.METAGROSS, Species.ALOLA_DUGTRIO],
-  OLIVIA: [Species.RELICANTH, Species.CARBINK, Species.ALOLA_GOLEM, Species.LYCANROC],
-  ACEROLA: [[Species.BANETTE, Species.DRIFBLIM], Species.MIMIKYU, Species.DHELMISE, Species.PALOSSAND],
-  KAHILI: [[Species.BRAVIARY, Species.MANDIBUZZ], Species.HAWLUCHA, Species.ORICORIO, Species.TOUCANNON],
-  MARNIE_ELITE: [Species.MORPEKO, Species.LIEPARD, [Species.TOXICROAK, Species.SCRAFTY], Species.GRIMMSNARL],
-  NESSA_ELITE: [Species.GOLISOPOD, [Species.PELIPPER, Species.QUAGSIRE], Species.TOXAPEX, Species.DREDNAW],
-  BEA_ELITE: [Species.HAWLUCHA, [Species.GRAPPLOCT, Species.SIRFETCHD], Species.FALINKS, Species.MACHAMP],
-  ALLISTER_ELITE: [Species.DUSKNOIR, [Species.POLTEAGEIST, Species.RUNERIGUS], Species.CURSOLA, Species.GENGAR],
-  RAIHAN_ELITE: [Species.GOODRA, [Species.TORKOAL, Species.TURTONATOR], Species.FLYGON, Species.ARCHALUDON],
-  RIKA: [Species.WHISCASH, [Species.DONPHAN, Species.DUGTRIO], Species.CAMERUPT, Species.CLODSIRE],
-  POPPY: [Species.COPPERAJAH, Species.BRONZONG, Species.CORVIKNIGHT, Species.TINKATON],
-  LARRY_ELITE: [Species.STARAPTOR, Species.FLAMIGO, Species.ALTARIA, Species.TROPIUS],
-  HASSEL: [Species.NOIVERN, [Species.FLAPPLE, Species.APPLETUN], Species.DRAGALGE, Species.BAXCALIBUR],
-  CRISPIN: [Species.TALONFLAME, Species.CAMERUPT, Species.MAGMORTAR, Species.BLAZIKEN],
-  AMARYS: [Species.SKARMORY, Species.EMPOLEON, Species.SCIZOR, Species.METAGROSS],
-  LACEY: [Species.EXCADRILL, Species.PRIMARINA, [Species.ALCREMIE, Species.GRANBULL], Species.WHIMSICOTT],
-  DRAYTON: [Species.DRAGONITE, Species.ARCHALUDON, Species.HAXORUS, Species.SCEPTILE],
-  BLUE: [
-    [Species.GYARADOS, Species.EXEGGUTOR, Species.ARCANINE],
-    Species.HO_OH,
-    [Species.RHYPERIOR, Species.MAGNEZONE],
-  ], // Alakazam lead, Mega Pidgeot
-  RED: [Species.LUGIA, Species.SNORLAX, [Species.ESPEON, Species.UMBREON, Species.SYLVEON]], // GMax Pikachu lead, Mega gen 1 starter
-  LANCE_CHAMPION: [Species.DRAGONITE, Species.KINGDRA, Species.ALOLA_EXEGGUTOR], // Aerodactyl lead, Mega Latias/Latios
-  STEVEN: [Species.AGGRON, [Species.ARMALDO, Species.CRADILY], Species.DIALGA], // Skarmory lead, Mega Metagross
-  WALLACE: [Species.MILOTIC, Species.PALKIA, Species.LUDICOLO], // Pelipper lead, Mega Swampert
-  CYNTHIA: [Species.GIRATINA, Species.LUCARIO, Species.TOGEKISS], // Spiritomb lead, Mega Garchomp
-  ALDER: [Species.VOLCARONA, Species.ZEKROM, [Species.ACCELGOR, Species.ESCAVALIER], Species.KELDEO], // Bouffalant/Braviary lead
-  IRIS: [Species.HAXORUS, Species.RESHIRAM, Species.ARCHEOPS], // Druddigon lead, Gmax Lapras
-  DIANTHA: [Species.HAWLUCHA, Species.XERNEAS, Species.GOODRA], // Gourgeist lead, Mega Gardevoir
-  HAU: [
-    [Species.SOLGALEO, Species.LUNALA],
-    Species.NOIVERN,
-    [Species.DECIDUEYE, Species.INCINEROAR, Species.PRIMARINA],
-    [Species.TAPU_BULU, Species.TAPU_FINI, Species.TAPU_KOKO, Species.TAPU_LELE],
-  ], // Alola Raichu lead
-  LEON: [Species.DRAGAPULT, Species.ZACIAN, Species.AEGISLASH], // Rillaboom/Cinderace/Inteleon lead, GMax Charizard
-  GEETA: [
-    Species.MIRAIDON,
-    [Species.ESPATHRA, Species.VELUZA],
-    [Species.AVALUGG, Species.HISUI_AVALUGG],
-    Species.KINGAMBIT,
-  ], // Glimmora lead
-  NEMONA: [
-    Species.KORAIDON,
-    Species.PAWMOT,
-    [Species.DUDUNSPARCE, Species.ORTHWORM],
-    [Species.MEOWSCARADA, Species.SKELEDIRGE, Species.QUAQUAVAL],
-  ], // Lycanroc lead
-  KIERAN: [
-    [Species.GRIMMSNARL, Species.INCINEROAR, Species.PORYGON_Z],
-    Species.OGERPON,
-    Species.TERAPAGOS,
-    Species.HYDRAPPLE,
-  ], // Poliwrath/Politoed lead
-};
-
 export const trainerConfigs: TrainerConfigs = {
-  [TrainerType.UNKNOWN]: new TrainerConfig(0).setHasGenders(),
-  [TrainerType.ACE_TRAINER]: new TrainerConfig(++t)
-    .setHasGenders("Ace Trainer Female")
-    .setHasDouble("Ace Duo")
-    .setMoneyMultiplier(2.25)
-    .setEncounterBgm(TrainerType.ACE_TRAINER)
-    .setPartyTemplateFunc((scene) =>
-      getWavePartyTemplate(
-        scene,
-        trainerPartyTemplates.THREE_WEAK_BALANCED,
-        trainerPartyTemplates.FOUR_WEAK_BALANCED,
-        trainerPartyTemplates.FIVE_WEAK_BALANCED,
-        trainerPartyTemplates.SIX_WEAK_BALANCED,
-      ),
-    ),
-  [TrainerType.ARTIST]: new TrainerConfig(++t)
-    .setEncounterBgm(TrainerType.RICH)
-    .setPartyTemplates(trainerPartyTemplates.ONE_STRONG, trainerPartyTemplates.TWO_AVG, trainerPartyTemplates.THREE_AVG)
-    .setSpeciesPools([Species.SMEARGLE]),
-  [TrainerType.BACKERS]: new TrainerConfig(++t)
-    .setHasGenders("Backers")
-    .setDoubleOnly()
-    .setEncounterBgm(TrainerType.CYCLIST),
-  [TrainerType.BACKPACKER]: new TrainerConfig(++t)
-    .setHasGenders("Backpacker Female")
-    .setHasDouble("Backpackers")
-    .setSpeciesFilter((s) => s.isOfType(Type.FLYING) || s.isOfType(Type.ROCK))
-    .setEncounterBgm(TrainerType.BACKPACKER)
-    .setPartyTemplates(
-      trainerPartyTemplates.ONE_STRONG,
-      trainerPartyTemplates.ONE_WEAK_ONE_STRONG,
-      trainerPartyTemplates.ONE_AVG_ONE_STRONG,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.RHYHORN,
-        Species.AIPOM,
-        Species.MAKUHITA,
-        Species.MAWILE,
-        Species.NUMEL,
-        Species.LILLIPUP,
-        Species.SANDILE,
-        Species.WOOLOO,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.GIRAFARIG,
-        Species.ZANGOOSE,
-        Species.SEVIPER,
-        Species.CUBCHOO,
-        Species.PANCHAM,
-        Species.SKIDDO,
-        Species.MUDBRAY,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.TAUROS,
-        Species.STANTLER,
-        Species.DARUMAKA,
-        Species.BOUFFALANT,
-        Species.DEERLING,
-        Species.IMPIDIMP,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.GALAR_DARUMAKA, Species.TEDDIURSA],
-    }),
-  [TrainerType.BAKER]: new TrainerConfig(++t)
-    .setEncounterBgm(TrainerType.CLERK)
-    .setMoneyMultiplier(1.35)
-    .setSpeciesFilter((s) => s.isOfType(Type.GRASS) || s.isOfType(Type.FIRE)),
-  [TrainerType.BEAUTY]: new TrainerConfig(++t).setMoneyMultiplier(1.55).setEncounterBgm(TrainerType.PARASOL_LADY),
-  [TrainerType.BIKER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.4)
-    .setEncounterBgm(TrainerType.ROUGHNECK)
-    .setSpeciesFilter((s) => s.isOfType(Type.POISON)),
-  [TrainerType.BLACK_BELT]: new TrainerConfig(++t)
-    .setHasGenders("Battle Girl", TrainerType.PSYCHIC)
-    .setHasDouble("Crush Kin")
-    .setEncounterBgm(TrainerType.ROUGHNECK)
-    .setSpecialtyTypes(Type.FIGHTING)
-    .setPartyTemplates(
-      trainerPartyTemplates.TWO_WEAK_ONE_AVG,
-      trainerPartyTemplates.TWO_WEAK_ONE_AVG,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.TWO_WEAK_ONE_STRONG,
-      trainerPartyTemplates.THREE_AVG,
-      trainerPartyTemplates.TWO_AVG_ONE_STRONG,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.NIDORAN_F,
-        Species.NIDORAN_M,
-        Species.MACHOP,
-        Species.MAKUHITA,
-        Species.MEDITITE,
-        Species.CROAGUNK,
-        Species.TIMBURR,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.MANKEY,
-        Species.POLIWRATH,
-        Species.TYROGUE,
-        Species.BRELOOM,
-        Species.SCRAGGY,
-        Species.MIENFOO,
-        Species.PANCHAM,
-        Species.STUFFUL,
-        Species.CRABRAWLER,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.HERACROSS,
-        Species.RIOLU,
-        Species.THROH,
-        Species.SAWK,
-        Species.PASSIMIAN,
-        Species.CLOBBOPUS,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [
-        Species.HITMONTOP,
-        Species.INFERNAPE,
-        Species.GALLADE,
-        Species.HAWLUCHA,
-        Species.HAKAMO_O,
-      ],
-      [TrainerPoolTier.ULTRA_RARE]: [Species.KUBFU],
-    }),
-  [TrainerType.BREEDER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.325)
-    .setEncounterBgm(TrainerType.POKEFAN)
-    .setHasGenders("Breeder Female")
-    .setHasDouble("Breeders")
-    .setPartyTemplateFunc((scene) =>
-      getWavePartyTemplate(
-        scene,
-        trainerPartyTemplates.FOUR_WEAKER,
-        trainerPartyTemplates.FIVE_WEAKER,
-        trainerPartyTemplates.SIX_WEAKER,
-      ),
-    )
-    .setSpeciesFilter((s) => s.baseTotal < 450),
-  [TrainerType.CLERK]: new TrainerConfig(++t)
-    .setHasGenders("Clerk Female")
-    .setHasDouble("Colleagues")
-    .setEncounterBgm(TrainerType.CLERK)
-    .setPartyTemplates(
-      trainerPartyTemplates.TWO_WEAK,
-      trainerPartyTemplates.THREE_WEAK,
-      trainerPartyTemplates.ONE_AVG,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.TWO_WEAK_ONE_AVG,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.MEOWTH,
-        Species.PSYDUCK,
-        Species.BUDEW,
-        Species.PIDOVE,
-        Species.CINCCINO,
-        Species.LITLEO,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.JIGGLYPUFF,
-        Species.MAGNEMITE,
-        Species.MARILL,
-        Species.COTTONEE,
-        Species.SKIDDO,
-      ],
-      [TrainerPoolTier.RARE]: [Species.BUIZEL, Species.SNEASEL, Species.KLEFKI, Species.INDEEDEE],
-    }),
-  [TrainerType.CYCLIST]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.3)
-    .setHasGenders("Cyclist Female")
-    .setHasDouble("Cyclists")
-    .setEncounterBgm(TrainerType.CYCLIST)
-    .setPartyTemplates(trainerPartyTemplates.TWO_WEAK, trainerPartyTemplates.ONE_AVG)
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [Species.PICHU, Species.STARLY, Species.TAILLOW, Species.BOLTUND],
-      [TrainerPoolTier.UNCOMMON]: [Species.DODUO, Species.ELECTRIKE, Species.BLITZLE, Species.WATTREL],
-      [TrainerPoolTier.RARE]: [Species.YANMA, Species.NINJASK, Species.WHIRLIPEDE, Species.EMOLGA],
-      [TrainerPoolTier.SUPER_RARE]: [Species.ACCELGOR, Species.DREEPY],
-    }),
-  [TrainerType.DANCER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.55)
-    .setEncounterBgm(TrainerType.CYCLIST)
-    .setPartyTemplates(
-      trainerPartyTemplates.TWO_WEAK,
-      trainerPartyTemplates.ONE_AVG,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.TWO_WEAK_SAME_TWO_WEAK_SAME,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [Species.RALTS, Species.SPOINK, Species.LOTAD, Species.BUDEW],
-      [TrainerPoolTier.UNCOMMON]: [Species.SPINDA, Species.SWABLU, Species.MARACTUS],
-      [TrainerPoolTier.RARE]: [Species.BELLOSSOM, Species.HITMONTOP, Species.MIME_JR, Species.ORICORIO],
-      [TrainerPoolTier.SUPER_RARE]: [Species.POPPLIO],
-    }),
-  [TrainerType.DEPOT_AGENT]: new TrainerConfig(++t).setMoneyMultiplier(1.45).setEncounterBgm(TrainerType.CLERK),
-  [TrainerType.DOCTOR]: new TrainerConfig(++t)
-    .setHasGenders("Nurse", "lass")
-    .setHasDouble("Medical Team")
-    .setMoneyMultiplier(3)
-    .setEncounterBgm(TrainerType.CLERK)
-    .setSpeciesFilter((s) => !!s.getLevelMoves().find((plm) => plm[1] === Moves.HEAL_PULSE)),
-  [TrainerType.FIREBREATHER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.4)
-    .setEncounterBgm(TrainerType.ROUGHNECK)
-    .setSpeciesFilter((s) => !!s.getLevelMoves().find((plm) => plm[1] === Moves.SMOG) || s.isOfType(Type.FIRE)),
-  [TrainerType.FISHERMAN]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.25)
-    .setEncounterBgm(TrainerType.BACKPACKER)
-    .setSpecialtyTypes(Type.WATER)
-    .setPartyTemplates(
-      trainerPartyTemplates.TWO_WEAK_SAME_ONE_AVG,
-      trainerPartyTemplates.ONE_AVG,
-      trainerPartyTemplates.THREE_WEAK_SAME,
-      trainerPartyTemplates.ONE_STRONG,
-      trainerPartyTemplates.SIX_WEAKER,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.TENTACOOL,
-        Species.MAGIKARP,
-        Species.GOLDEEN,
-        Species.STARYU,
-        Species.REMORAID,
-        Species.SKRELP,
-        Species.CLAUNCHER,
-        Species.ARROKUDA,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.POLIWAG,
-        Species.SHELLDER,
-        Species.KRABBY,
-        Species.HORSEA,
-        Species.CARVANHA,
-        Species.BARBOACH,
-        Species.CORPHISH,
-        Species.FINNEON,
-        Species.TYMPOLE,
-        Species.BASCULIN,
-        Species.FRILLISH,
-        Species.INKAY,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.CHINCHOU,
-        Species.CORSOLA,
-        Species.WAILMER,
-        Species.BARBOACH,
-        Species.CLAMPERL,
-        Species.LUVDISC,
-        Species.MANTYKE,
-        Species.ALOMOMOLA,
-        Species.TATSUGIRI,
-        Species.VELUZA,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.LAPRAS, Species.FEEBAS, Species.RELICANTH, Species.DONDOZO],
-    }),
-  [TrainerType.GUITARIST]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.2)
-    .setEncounterBgm(TrainerType.ROUGHNECK)
-    .setSpecialtyTypes(Type.ELECTRIC)
-    .setSpeciesFilter((s) => s.isOfType(Type.ELECTRIC)),
-  [TrainerType.HARLEQUIN]: new TrainerConfig(++t)
-    .setEncounterBgm(TrainerType.PSYCHIC)
-    .setSpeciesFilter((s) => tmSpecies[Moves.TRICK_ROOM].indexOf(s.speciesId) > -1),
-  [TrainerType.HIKER]: new TrainerConfig(++t)
-    .setEncounterBgm(TrainerType.BACKPACKER)
-    .setPartyTemplates(
-      trainerPartyTemplates.TWO_AVG_SAME_ONE_AVG,
-      trainerPartyTemplates.TWO_AVG_SAME_ONE_STRONG,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.FOUR_WEAK,
-      trainerPartyTemplates.ONE_STRONG,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.SANDSHREW,
-        Species.DIGLETT,
-        Species.GEODUDE,
-        Species.MACHOP,
-        Species.ARON,
-        Species.ROGGENROLA,
-        Species.DRILBUR,
-        Species.NACLI,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.ZUBAT,
-        Species.RHYHORN,
-        Species.ONIX,
-        Species.CUBONE,
-        Species.WOOBAT,
-        Species.SWINUB,
-        Species.NOSEPASS,
-        Species.HIPPOPOTAS,
-        Species.DWEBBLE,
-        Species.KLAWF,
-        Species.TOEDSCOOL,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.TORKOAL,
-        Species.TRAPINCH,
-        Species.BARBOACH,
-        Species.GOLETT,
-        Species.ALOLA_DIGLETT,
-        Species.ALOLA_GEODUDE,
-        Species.GALAR_STUNFISK,
-        Species.PALDEA_WOOPER,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.MAGBY, Species.LARVITAR],
-    }),
-  [TrainerType.HOOLIGANS]: new TrainerConfig(++t)
-    .setDoubleOnly()
-    .setEncounterBgm(TrainerType.ROUGHNECK)
-    .setSpeciesFilter((s) => s.isOfType(Type.POISON) || s.isOfType(Type.DARK)),
-  [TrainerType.HOOPSTER]: new TrainerConfig(++t).setMoneyMultiplier(1.2).setEncounterBgm(TrainerType.CYCLIST),
-  [TrainerType.INFIELDER]: new TrainerConfig(++t).setMoneyMultiplier(1.2).setEncounterBgm(TrainerType.CYCLIST),
-  [TrainerType.JANITOR]: new TrainerConfig(++t).setMoneyMultiplier(1.1).setEncounterBgm(TrainerType.CLERK),
-  [TrainerType.LINEBACKER]: new TrainerConfig(++t).setMoneyMultiplier(1.2).setEncounterBgm(TrainerType.CYCLIST),
-  [TrainerType.MAID]: new TrainerConfig(++t).setMoneyMultiplier(1.6).setEncounterBgm(TrainerType.RICH),
-  [TrainerType.MUSICIAN]: new TrainerConfig(++t)
-    .setEncounterBgm(TrainerType.ROUGHNECK)
-    .setSpeciesFilter((s) => !!s.getLevelMoves().find((plm) => plm[1] === Moves.SING)),
-  [TrainerType.HEX_MANIAC]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .setEncounterBgm(TrainerType.PSYCHIC)
-    .setPartyTemplates(
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.ONE_AVG_ONE_STRONG,
-      trainerPartyTemplates.TWO_AVG_SAME_ONE_AVG,
-      trainerPartyTemplates.THREE_AVG,
-      trainerPartyTemplates.TWO_STRONG,
-    )
-    .setSpeciesFilter((s) => s.isOfType(Type.GHOST)),
-  [TrainerType.NURSERY_AIDE]: new TrainerConfig(++t).setMoneyMultiplier(1.3).setEncounterBgm("lass"),
-  [TrainerType.OFFICER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.55)
-    .setEncounterBgm(TrainerType.CLERK)
-    .setPartyTemplates(
-      trainerPartyTemplates.ONE_AVG,
-      trainerPartyTemplates.ONE_STRONG,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.TWO_WEAK_SAME_ONE_AVG,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.VULPIX,
-        Species.GROWLITHE,
-        Species.SNUBBULL,
-        Species.POOCHYENA,
-        Species.ELECTRIKE,
-        Species.LILLIPUP,
-        Species.YAMPER,
-        Species.FIDOUGH,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [Species.HOUNDOUR, Species.ROCKRUFF, Species.MASCHIFF],
-      [TrainerPoolTier.RARE]: [Species.JOLTEON, Species.RIOLU],
-      [TrainerPoolTier.SUPER_RARE]: [],
-      [TrainerPoolTier.ULTRA_RARE]: [Species.ENTEI, Species.SUICUNE, Species.RAIKOU],
-    }),
-  [TrainerType.PARASOL_LADY]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.55)
-    .setEncounterBgm(TrainerType.PARASOL_LADY)
-    .setSpeciesFilter((s) => s.isOfType(Type.WATER)),
-  [TrainerType.PILOT]: new TrainerConfig(++t)
-    .setEncounterBgm(TrainerType.CLERK)
-    .setSpeciesFilter((s) => tmSpecies[Moves.FLY].indexOf(s.speciesId) > -1),
-  [TrainerType.POKEFAN]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.4)
-    .setName("PokéFan")
-    .setHasGenders("PokéFan Female")
-    .setHasDouble("PokéFan Family")
-    .setEncounterBgm(TrainerType.POKEFAN)
-    .setPartyTemplates(
-      trainerPartyTemplates.SIX_WEAKER,
-      trainerPartyTemplates.FOUR_WEAK,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.ONE_STRONG,
-      trainerPartyTemplates.FOUR_WEAK_SAME,
-      trainerPartyTemplates.FIVE_WEAK,
-      trainerPartyTemplates.SIX_WEAKER_SAME,
-    ),
-  [TrainerType.PRESCHOOLER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(0.2)
-    .setEncounterBgm(TrainerType.YOUNGSTER)
-    .setHasGenders("Preschooler Female", "lass")
-    .setHasDouble("Preschoolers")
-    .setPartyTemplates(
-      trainerPartyTemplates.THREE_WEAK,
-      trainerPartyTemplates.FOUR_WEAKER,
-      trainerPartyTemplates.TWO_WEAK_SAME_ONE_AVG,
-      trainerPartyTemplates.FIVE_WEAKER,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.CATERPIE,
-        Species.PICHU,
-        Species.SANDSHREW,
-        Species.LEDYBA,
-        Species.BUDEW,
-        Species.BURMY,
-        Species.WOOLOO,
-        Species.PAWMI,
-        Species.SMOLIV,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.EEVEE,
-        Species.CLEFFA,
-        Species.IGGLYBUFF,
-        Species.SWINUB,
-        Species.WOOPER,
-        Species.DRIFLOON,
-        Species.DEDENNE,
-        Species.STUFFUL,
-      ],
-      [TrainerPoolTier.RARE]: [Species.RALTS, Species.RIOLU, Species.JOLTIK, Species.TANDEMAUS],
-      [TrainerPoolTier.SUPER_RARE]: [Species.DARUMAKA, Species.TINKATINK],
-    }),
-  [TrainerType.PSYCHIC]: new TrainerConfig(++t)
-    .setHasGenders("Psychic Female")
-    .setHasDouble("Psychics")
-    .setMoneyMultiplier(1.4)
-    .setEncounterBgm(TrainerType.PSYCHIC)
-    .setPartyTemplates(
-      trainerPartyTemplates.TWO_WEAK,
-      trainerPartyTemplates.TWO_AVG,
-      trainerPartyTemplates.TWO_WEAK_SAME_ONE_AVG,
-      trainerPartyTemplates.TWO_WEAK_SAME_TWO_WEAK_SAME,
-      trainerPartyTemplates.ONE_STRONGER,
-    )
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.ABRA,
-        Species.DROWZEE,
-        Species.RALTS,
-        Species.SPOINK,
-        Species.GOTHITA,
-        Species.SOLOSIS,
-        Species.BLIPBUG,
-        Species.ESPURR,
-        Species.HATENNA,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.MIME_JR,
-        Species.EXEGGCUTE,
-        Species.MEDITITE,
-        Species.NATU,
-        Species.EXEGGCUTE,
-        Species.WOOBAT,
-        Species.INKAY,
-        Species.ORANGURU,
-      ],
-      [TrainerPoolTier.RARE]: [Species.ELGYEM, Species.SIGILYPH, Species.BALTOY, Species.GIRAFARIG, Species.MEOWSTIC],
-      [TrainerPoolTier.SUPER_RARE]: [Species.BELDUM, Species.ESPEON, Species.STANTLER],
-    }),
-  [TrainerType.RANGER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.4)
-    .setName("Pokémon Ranger")
-    .setEncounterBgm(TrainerType.BACKPACKER)
-    .setHasGenders("Pokémon Ranger Female")
-    .setHasDouble("Pokémon Rangers")
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.PICHU,
-        Species.GROWLITHE,
-        Species.PONYTA,
-        Species.ZIGZAGOON,
-        Species.SEEDOT,
-        Species.BIDOOF,
-        Species.RIOLU,
-        Species.SEWADDLE,
-        Species.SKIDDO,
-        Species.SALANDIT,
-        Species.YAMPER,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.AZURILL,
-        Species.TAUROS,
-        Species.MAREEP,
-        Species.FARFETCHD,
-        Species.TEDDIURSA,
-        Species.SHROOMISH,
-        Species.ELECTRIKE,
-        Species.BUDEW,
-        Species.BUIZEL,
-        Species.MUDBRAY,
-        Species.STUFFUL,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.EEVEE,
-        Species.SCYTHER,
-        Species.KANGASKHAN,
-        Species.RALTS,
-        Species.MUNCHLAX,
-        Species.ZORUA,
-        Species.PALDEA_TAUROS,
-        Species.TINKATINK,
-        Species.CYCLIZAR,
-        Species.FLAMIGO,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.LARVESTA],
-    }),
-  [TrainerType.RICH]: new TrainerConfig(++t)
-    .setMoneyMultiplier(5)
-    .setName("Gentleman")
-    .setHasGenders("Madame")
-    .setHasDouble("Rich Couple"),
-  [TrainerType.RICH_KID]: new TrainerConfig(++t)
-    .setMoneyMultiplier(3.75)
-    .setName("Rich Boy")
-    .setHasGenders("Lady")
-    .setHasDouble("Rich Kids")
-    .setEncounterBgm(TrainerType.RICH),
-  [TrainerType.ROUGHNECK]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.4)
-    .setEncounterBgm(TrainerType.ROUGHNECK)
-    .setSpeciesFilter((s) => s.isOfType(Type.DARK)),
-  [TrainerType.SAILOR]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.4)
-    .setEncounterBgm(TrainerType.BACKPACKER)
-    .setSpeciesFilter((s) => s.isOfType(Type.WATER) || s.isOfType(Type.FIGHTING)),
-  [TrainerType.SCIENTIST]: new TrainerConfig(++t)
-    .setHasGenders("Scientist Female")
-    .setHasDouble("Scientists")
-    .setMoneyMultiplier(1.7)
-    .setEncounterBgm(TrainerType.SCIENTIST)
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [Species.MAGNEMITE, Species.GRIMER, Species.DROWZEE, Species.VOLTORB, Species.KOFFING],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.BALTOY,
-        Species.BRONZOR,
-        Species.FERROSEED,
-        Species.KLINK,
-        Species.CHARJABUG,
-        Species.BLIPBUG,
-        Species.HELIOPTILE,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.ABRA,
-        Species.DITTO,
-        Species.PORYGON,
-        Species.ELEKID,
-        Species.SOLOSIS,
-        Species.GALAR_WEEZING,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [
-        Species.OMANYTE,
-        Species.KABUTO,
-        Species.AERODACTYL,
-        Species.LILEEP,
-        Species.ANORITH,
-        Species.CRANIDOS,
-        Species.SHIELDON,
-        Species.TIRTOUGA,
-        Species.ARCHEN,
-        Species.ARCTOVISH,
-        Species.ARCTOZOLT,
-        Species.DRACOVISH,
-        Species.DRACOZOLT,
-      ],
-      [TrainerPoolTier.ULTRA_RARE]: [Species.ROTOM, Species.MELTAN],
-    }),
-  [TrainerType.SMASHER]: new TrainerConfig(++t).setMoneyMultiplier(1.2).setEncounterBgm(TrainerType.CYCLIST),
-  [TrainerType.SNOW_WORKER]: new TrainerConfig(++t)
-    .setName("Worker")
-    .setHasDouble("Workers")
-    .setMoneyMultiplier(1.7)
-    .setEncounterBgm(TrainerType.CLERK)
-    .setSpeciesFilter((s) => s.isOfType(Type.ICE) || s.isOfType(Type.STEEL)),
-  [TrainerType.STRIKER]: new TrainerConfig(++t).setMoneyMultiplier(1.2).setEncounterBgm(TrainerType.CYCLIST),
-  [TrainerType.SCHOOL_KID]: new TrainerConfig(++t)
-    .setMoneyMultiplier(0.75)
-    .setEncounterBgm(TrainerType.YOUNGSTER)
-    .setHasGenders("School Kid Female", "lass")
-    .setHasDouble("School Kids")
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.ODDISH,
-        Species.EXEGGCUTE,
-        Species.TEDDIURSA,
-        Species.WURMPLE,
-        Species.RALTS,
-        Species.SHROOMISH,
-        Species.FLETCHLING,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [Species.VOLTORB, Species.WHISMUR, Species.MEDITITE, Species.MIME_JR, Species.NYMBLE],
-      [TrainerPoolTier.RARE]: [Species.TANGELA, Species.EEVEE, Species.YANMA],
-      [TrainerPoolTier.SUPER_RARE]: [Species.TADBULB],
-    }),
-  [TrainerType.SWIMMER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.3)
-    .setEncounterBgm(TrainerType.PARASOL_LADY)
-    .setHasGenders("Swimmer Female")
-    .setHasDouble("Swimmers")
-    .setSpecialtyTypes(Type.WATER)
-    .setSpeciesFilter((s) => s.isOfType(Type.WATER)),
-  [TrainerType.TWINS]: new TrainerConfig(++t)
-    .setDoubleOnly()
-    .setMoneyMultiplier(0.65)
-    .setUseSameSeedForAllMembers()
-    .setPartyTemplateFunc((scene) =>
-      getWavePartyTemplate(
-        scene,
-        trainerPartyTemplates.TWO_WEAK,
-        trainerPartyTemplates.TWO_AVG,
-        trainerPartyTemplates.TWO_STRONG,
-      ),
-    )
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([
-        Species.PLUSLE,
-        Species.VOLBEAT,
-        Species.PACHIRISU,
-        Species.SILCOON,
-        Species.METAPOD,
-        Species.IGGLYBUFF,
-        Species.PETILIL,
-        Species.EEVEE,
-      ]),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [
-          Species.MINUN,
-          Species.ILLUMISE,
-          Species.EMOLGA,
-          Species.CASCOON,
-          Species.KAKUNA,
-          Species.CLEFFA,
-          Species.COTTONEE,
-          Species.EEVEE,
-        ],
-        TrainerSlot.TRAINER_PARTNER,
-      ),
-    )
-    .setEncounterBgm(TrainerType.TWINS),
-  [TrainerType.VETERAN]: new TrainerConfig(++t)
-    .setHasGenders("Veteran Female")
-    .setHasDouble("Veteran Duo")
-    .setMoneyMultiplier(2.5)
-    .setEncounterBgm(TrainerType.ACE_TRAINER)
-    .setSpeciesFilter((s) => s.isOfType(Type.DRAGON)),
-  [TrainerType.WAITER]: new TrainerConfig(++t)
-    .setHasGenders("Waitress")
-    .setHasDouble("Restaurant Staff")
-    .setMoneyMultiplier(1.5)
-    .setEncounterBgm(TrainerType.CLERK)
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.CLEFFA,
-        Species.CHATOT,
-        Species.PANSAGE,
-        Species.PANSEAR,
-        Species.PANPOUR,
-        Species.MINCCINO,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [Species.TROPIUS, Species.PETILIL, Species.BOUNSWEET, Species.INDEEDEE],
-      [TrainerPoolTier.RARE]: [Species.APPLIN, Species.SINISTEA, Species.POLTCHAGEIST],
-    }),
-  [TrainerType.WORKER]: new TrainerConfig(++t)
-    .setHasGenders("Worker Female")
-    .setHasDouble("Workers")
-    .setEncounterBgm(TrainerType.CLERK)
-    .setMoneyMultiplier(1.7)
-    .setSpeciesFilter((s) => s.isOfType(Type.ROCK) || s.isOfType(Type.STEEL)),
-  [TrainerType.YOUNGSTER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(0.5)
-    .setEncounterBgm(TrainerType.YOUNGSTER)
-    .setHasGenders("Lass", "lass")
-    .setHasDouble("Beginners")
-    .setPartyTemplates(trainerPartyTemplates.TWO_WEAKER)
-    .setSpeciesPools([
-      Species.CATERPIE,
-      Species.WEEDLE,
-      Species.RATTATA,
-      Species.SENTRET,
-      Species.POOCHYENA,
-      Species.ZIGZAGOON,
-      Species.WURMPLE,
-      Species.BIDOOF,
-      Species.PATRAT,
-      Species.LILLIPUP,
-    ]),
-  [TrainerType.ROCKET_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Rocket Grunt Female")
-    .setHasDouble("Rocket Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_rocket_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.WEEDLE,
-        Species.RATTATA,
-        Species.EKANS,
-        Species.SANDSHREW,
-        Species.ZUBAT,
-        Species.GEODUDE,
-        Species.KOFFING,
-        Species.GRIMER,
-        Species.ODDISH,
-        Species.SLOWPOKE,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.GYARADOS,
-        Species.LICKITUNG,
-        Species.TAUROS,
-        Species.MANKEY,
-        Species.SCYTHER,
-        Species.ELEKID,
-        Species.MAGBY,
-        Species.CUBONE,
-        Species.GROWLITHE,
-        Species.MURKROW,
-        Species.GASTLY,
-        Species.EXEGGCUTE,
-        Species.VOLTORB,
-        Species.MAGNEMITE,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.PORYGON,
-        Species.ALOLA_RATTATA,
-        Species.ALOLA_SANDSHREW,
-        Species.ALOLA_MEOWTH,
-        Species.ALOLA_GRIMER,
-        Species.ALOLA_GEODUDE,
-        Species.PALDEA_TAUROS,
-        Species.OMANYTE,
-        Species.KABUTO,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.DRATINI, Species.LARVITAR],
-    }),
-  [TrainerType.ARCHER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("rocket_admin", "rocket", [Species.HOUNDOOM])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_rocket_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.ARIANA]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("rocket_admin_female", "rocket", [Species.ARBOK])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_rocket_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.PROTON]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("rocket_admin", "rocket", [Species.CROBAT])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_rocket_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.PETREL]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("rocket_admin", "rocket", [Species.WEEZING])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_rocket_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.MAGMA_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Magma Grunt Female")
-    .setHasDouble("Magma Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aqua_magma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.SLUGMA,
-        Species.POOCHYENA,
-        Species.NUMEL,
-        Species.ZIGZAGOON,
-        Species.DIGLETT,
-        Species.MAGBY,
-        Species.TORKOAL,
-        Species.GROWLITHE,
-        Species.BALTOY,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.SOLROCK,
-        Species.HIPPOPOTAS,
-        Species.SANDACONDA,
-        Species.PHANPY,
-        Species.ROLYCOLY,
-        Species.GLIGAR,
-        Species.RHYHORN,
-        Species.HEATMOR,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.TRAPINCH,
-        Species.LILEEP,
-        Species.ANORITH,
-        Species.HISUI_GROWLITHE,
-        Species.TURTONATOR,
-        Species.ARON,
-        Species.TOEDSCOOL,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.CAPSAKID, Species.CHARCADET],
-    }),
-  [TrainerType.TABITHA]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("magma_admin", "magma", [Species.CAMERUPT])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aqua_magma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.COURTNEY]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("magma_admin_female", "magma", [Species.CAMERUPT])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aqua_magma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.AQUA_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Aqua Grunt Female")
-    .setHasDouble("Aqua Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aqua_magma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.CARVANHA,
-        Species.WAILMER,
-        Species.ZIGZAGOON,
-        Species.LOTAD,
-        Species.CORPHISH,
-        Species.SPHEAL,
-        Species.REMORAID,
-        Species.QWILFISH,
-        Species.BARBOACH,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.CLAMPERL,
-        Species.CHINCHOU,
-        Species.WOOPER,
-        Species.WINGULL,
-        Species.TENTACOOL,
-        Species.AZURILL,
-        Species.CLOBBOPUS,
-        Species.HORSEA,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.MANTYKE,
-        Species.DHELMISE,
-        Species.HISUI_QWILFISH,
-        Species.ARROKUDA,
-        Species.PALDEA_WOOPER,
-        Species.SKRELP,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.DONDOZO, Species.BASCULEGION],
-    }),
-  [TrainerType.MATT]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("aqua_admin", "aqua", [Species.SHARPEDO])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aqua_magma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.SHELLY]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("aqua_admin_female", "aqua", [Species.SHARPEDO])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aqua_magma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.GALACTIC_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Galactic Grunt Female")
-    .setHasDouble("Galactic Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_galactic_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.GLAMEOW,
-        Species.STUNKY,
-        Species.CROAGUNK,
-        Species.SHINX,
-        Species.WURMPLE,
-        Species.BRONZOR,
-        Species.DRIFLOON,
-        Species.BURMY,
-        Species.CARNIVINE,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.LICKITUNG,
-        Species.RHYHORN,
-        Species.TANGELA,
-        Species.ZUBAT,
-        Species.YANMA,
-        Species.SKORUPI,
-        Species.GLIGAR,
-        Species.SWINUB,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.HISUI_GROWLITHE,
-        Species.HISUI_QWILFISH,
-        Species.SNEASEL,
-        Species.ELEKID,
-        Species.MAGBY,
-        Species.DUSKULL,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.ROTOM, Species.SPIRITOMB, Species.HISUI_SNEASEL],
-    }),
-  [TrainerType.JUPITER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("galactic_commander_female", "galactic", [Species.SKUNTANK])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_galactic_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.MARS]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("galactic_commander_female", "galactic", [Species.PURUGLY])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_galactic_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.SATURN]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("galactic_commander", "galactic", [Species.TOXICROAK])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_galactic_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.PLASMA_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Plasma Grunt Female")
-    .setHasDouble("Plasma Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_plasma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.PATRAT,
-        Species.LILLIPUP,
-        Species.PURRLOIN,
-        Species.SCRAFTY,
-        Species.WOOBAT,
-        Species.VANILLITE,
-        Species.SANDILE,
-        Species.TRUBBISH,
-        Species.TYMPOLE,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.FRILLISH,
-        Species.VENIPEDE,
-        Species.GOLETT,
-        Species.TIMBURR,
-        Species.DARUMAKA,
-        Species.FOONGUS,
-        Species.JOLTIK,
-        Species.CUBCHOO,
-        Species.KLINK,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.PAWNIARD,
-        Species.RUFFLET,
-        Species.VULLABY,
-        Species.ZORUA,
-        Species.DRILBUR,
-        Species.MIENFOO,
-        Species.DURANT,
-        Species.BOUFFALANT,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.DRUDDIGON, Species.HISUI_ZORUA, Species.AXEW, Species.DEINO],
-    }),
-  [TrainerType.ZINZOLIN]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("plasma_sage", "plasma", [Species.CRYOGONAL])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_plasma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.ROOD]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("plasma_sage", "plasma", [Species.SWOOBAT])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_plasma_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.FLARE_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Flare Grunt Female")
-    .setHasDouble("Flare Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_flare_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.FLETCHLING,
-        Species.LITLEO,
-        Species.PONYTA,
-        Species.INKAY,
-        Species.HOUNDOUR,
-        Species.SKORUPI,
-        Species.SCRAFTY,
-        Species.CROAGUNK,
-        Species.SCATTERBUG,
-        Species.ESPURR,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.HELIOPTILE,
-        Species.ELECTRIKE,
-        Species.SKRELP,
-        Species.PANCHAM,
-        Species.PURRLOIN,
-        Species.POOCHYENA,
-        Species.BINACLE,
-        Species.CLAUNCHER,
-        Species.PUMPKABOO,
-        Species.PHANTUMP,
-        Species.FOONGUS,
-      ],
-      [TrainerPoolTier.RARE]: [Species.LITWICK, Species.SNEASEL, Species.PAWNIARD, Species.SLIGGOO],
-      [TrainerPoolTier.SUPER_RARE]: [Species.NOIBAT, Species.HISUI_SLIGGOO, Species.HISUI_AVALUGG],
-    }),
-  [TrainerType.BRYONY]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("flare_admin_female", "flare", [Species.LIEPARD])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_flare_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.XEROSIC]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("flare_admin", "flare", [Species.MALAMAR])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_flare_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.AETHER_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Aether Grunt Female")
-    .setHasDouble("Aether Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aether_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.PIKIPEK,
-        Species.ROCKRUFF,
-        Species.ALOLA_DIGLETT,
-        Species.ALOLA_EXEGGUTOR,
-        Species.YUNGOOS,
-        Species.CORSOLA,
-        Species.ALOLA_GEODUDE,
-        Species.ALOLA_RAICHU,
-        Species.BOUNSWEET,
-        Species.LILLIPUP,
-        Species.KOMALA,
-        Species.MORELULL,
-        Species.COMFEY,
-        Species.TOGEDEMARU,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.POLIWAG,
-        Species.STUFFUL,
-        Species.ORANGURU,
-        Species.PASSIMIAN,
-        Species.BRUXISH,
-        Species.MINIOR,
-        Species.WISHIWASHI,
-        Species.ALOLA_SANDSHREW,
-        Species.ALOLA_VULPIX,
-        Species.CRABRAWLER,
-        Species.CUTIEFLY,
-        Species.ORICORIO,
-        Species.MUDBRAY,
-        Species.PYUKUMUKU,
-        Species.ALOLA_MAROWAK,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.GALAR_CORSOLA,
-        Species.TURTONATOR,
-        Species.MIMIKYU,
-        Species.MAGNEMITE,
-        Species.DRAMPA,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.JANGMO_O, Species.PORYGON],
-    }),
-  [TrainerType.FABA]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("aether_admin", "aether", [Species.HYPNO])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_aether_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.SKULL_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Skull Grunt Female")
-    .setHasDouble("Skull Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_skull_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.SALANDIT,
-        Species.ALOLA_RATTATA,
-        Species.EKANS,
-        Species.ALOLA_MEOWTH,
-        Species.SCRAGGY,
-        Species.KOFFING,
-        Species.ALOLA_GRIMER,
-        Species.MAREANIE,
-        Species.SPINARAK,
-        Species.TRUBBISH,
-        Species.DROWZEE,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.FOMANTIS,
-        Species.SABLEYE,
-        Species.SANDILE,
-        Species.HOUNDOUR,
-        Species.ALOLA_MAROWAK,
-        Species.GASTLY,
-        Species.PANCHAM,
-        Species.ZUBAT,
-        Species.VENIPEDE,
-        Species.VULLABY,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.SANDYGAST,
-        Species.PAWNIARD,
-        Species.MIMIKYU,
-        Species.DHELMISE,
-        Species.WISHIWASHI,
-        Species.NYMBLE,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.GRUBBIN, Species.DEWPIDER],
-    }),
-  [TrainerType.PLUMERIA]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("skull_admin", "skull", [Species.SALAZZLE])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_skull_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.MACRO_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Macro Grunt Female")
-    .setHasDouble("Macro Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_macro_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.CUFANT,
-        Species.GALAR_MEOWTH,
-        Species.KLINK,
-        Species.ROOKIDEE,
-        Species.CRAMORANT,
-        Species.GALAR_ZIGZAGOON,
-        Species.SKWOVET,
-        Species.STEELIX,
-        Species.MAWILE,
-        Species.FERROSEED,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.DRILBUR,
-        Species.MAGNEMITE,
-        Species.HATENNA,
-        Species.ARROKUDA,
-        Species.APPLIN,
-        Species.GALAR_PONYTA,
-        Species.GALAR_YAMASK,
-        Species.SINISTEA,
-        Species.RIOLU,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.FALINKS,
-        Species.BELDUM,
-        Species.GALAR_FARFETCHD,
-        Species.GALAR_MR_MIME,
-        Species.HONEDGE,
-        Species.SCIZOR,
-        Species.GALAR_DARUMAKA,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.DURALUDON, Species.DREEPY],
-    }),
-  [TrainerType.OLEANA]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("macro_admin", "macro", [Species.GARBODOR])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_oleana")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene)),
-  [TrainerType.STAR_GRUNT]: new TrainerConfig(++t)
-    .setHasGenders("Star Grunt Female")
-    .setHasDouble("Star Grunts")
-    .setMoneyMultiplier(1.0)
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_star_grunt")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setSpeciesPools({
-      [TrainerPoolTier.COMMON]: [
-        Species.DUNSPARCE,
-        Species.HOUNDOUR,
-        Species.AZURILL,
-        Species.GULPIN,
-        Species.FOONGUS,
-        Species.FLETCHLING,
-        Species.LITLEO,
-        Species.FLABEBE,
-        Species.CRABRAWLER,
-        Species.NYMBLE,
-        Species.PAWMI,
-        Species.FIDOUGH,
-        Species.SQUAWKABILLY,
-        Species.MASCHIFF,
-        Species.SHROODLE,
-        Species.KLAWF,
-        Species.WIGLETT,
-        Species.PALDEA_WOOPER,
-      ],
-      [TrainerPoolTier.UNCOMMON]: [
-        Species.KOFFING,
-        Species.EEVEE,
-        Species.GIRAFARIG,
-        Species.RALTS,
-        Species.TORKOAL,
-        Species.SEVIPER,
-        Species.SCRAGGY,
-        Species.ZORUA,
-        Species.MIMIKYU,
-        Species.IMPIDIMP,
-        Species.FALINKS,
-        Species.CAPSAKID,
-        Species.TINKATINK,
-        Species.BOMBIRDIER,
-        Species.CYCLIZAR,
-        Species.FLAMIGO,
-        Species.PALDEA_TAUROS,
-      ],
-      [TrainerPoolTier.RARE]: [
-        Species.MANKEY,
-        Species.PAWNIARD,
-        Species.CHARCADET,
-        Species.FLITTLE,
-        Species.VAROOM,
-        Species.ORTHWORM,
-      ],
-      [TrainerPoolTier.SUPER_RARE]: [Species.DONDOZO, Species.GIMMIGHOUL],
-    }),
-  [TrainerType.GIACOMO]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("star_admin", "star_1", [Species.KINGAMBIT])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_star_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.REVAVROOM], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // Segin Starmobile
-        p.moveset = [
-          new PokemonMove(Moves.WICKED_TORQUE),
-          new PokemonMove(Moves.SPIN_OUT),
-          new PokemonMove(Moves.SHIFT_GEAR),
-          new PokemonMove(Moves.HIGH_HORSEPOWER),
-        ];
-      }),
-    ),
-  [TrainerType.MELA]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("star_admin", "star_2", [Species.ARMAROUGE])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_star_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.REVAVROOM], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 2; // Schedar Starmobile
-        p.moveset = [
-          new PokemonMove(Moves.BLAZING_TORQUE),
-          new PokemonMove(Moves.SPIN_OUT),
-          new PokemonMove(Moves.SHIFT_GEAR),
-          new PokemonMove(Moves.HIGH_HORSEPOWER),
-        ];
-      }),
-    ),
-  [TrainerType.ATTICUS]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("star_admin", "star_3", [Species.REVAVROOM])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_star_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.REVAVROOM], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 3; // Navi Starmobile
-        p.moveset = [
-          new PokemonMove(Moves.NOXIOUS_TORQUE),
-          new PokemonMove(Moves.SPIN_OUT),
-          new PokemonMove(Moves.SHIFT_GEAR),
-          new PokemonMove(Moves.HIGH_HORSEPOWER),
-        ];
-      }),
-    ),
-  [TrainerType.ORTEGA]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("star_admin", "star_4", [Species.DACHSBUN])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_star_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.REVAVROOM], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 4; // Ruchbah Starmobile
-        p.moveset = [
-          new PokemonMove(Moves.MAGICAL_TORQUE),
-          new PokemonMove(Moves.SPIN_OUT),
-          new PokemonMove(Moves.SHIFT_GEAR),
-          new PokemonMove(Moves.HIGH_HORSEPOWER),
-        ];
-      }),
-    ),
-  [TrainerType.ERI]: new TrainerConfig(++t)
-    .setMoneyMultiplier(1.5)
-    .initForEvilTeamAdmin("star_admin", "star_5", [Species.ANNIHILAPE])
-    .setEncounterBgm(TrainerType.PLASMA_GRUNT)
-    .setBattleBgm("battle_plasma_grunt")
-    .setMixedBattleBgm("battle_star_admin")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyTemplateFunc((scene) => getEvilGruntPartyTemplate(scene))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.REVAVROOM], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 5; // Caph Starmobile
-        p.moveset = [
-          new PokemonMove(Moves.COMBAT_TORQUE),
-          new PokemonMove(Moves.SPIN_OUT),
-          new PokemonMove(Moves.SHIFT_GEAR),
-          new PokemonMove(Moves.HIGH_HORSEPOWER),
-        ];
-      }),
-    ),
-
-  [TrainerType.BROCK]: new TrainerConfig((t = TrainerType.BROCK))
-    .initForGymLeader(signatureSpecies["BROCK"], true, Type.ROCK)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.MISTY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["MISTY"], false, Type.WATER)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.LT_SURGE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["LT_SURGE"], true, Type.ELECTRIC)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.ERIKA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["ERIKA"], false, Type.GRASS)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.JANINE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["JANINE"], false, Type.POISON)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.SABRINA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["SABRINA"], false, Type.PSYCHIC)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.BLAINE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BLAINE"], true, Type.FIRE)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.GIOVANNI]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["GIOVANNI"], true, Type.DARK)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.FALKNER]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["FALKNER"], true, Type.FLYING)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.BUGSY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BUGSY"], true, Type.BUG)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.WHITNEY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["WHITNEY"], false, Type.NORMAL)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.MORTY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["MORTY"], true, Type.GHOST)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.CHUCK]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CHUCK"], true, Type.FIGHTING)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.JASMINE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["JASMINE"], false, Type.STEEL)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.PRYCE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["PRYCE"], true, Type.ICE)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.CLAIR]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CLAIR"], false, Type.DRAGON)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.ROXANNE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["ROXANNE"], false, Type.ROCK)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym"),
-  [TrainerType.BRAWLY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BRAWLY"], true, Type.FIGHTING)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym"),
-  [TrainerType.WATTSON]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["WATTSON"], true, Type.ELECTRIC)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym"),
-  [TrainerType.FLANNERY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["FLANNERY"], false, Type.FIRE)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym"),
-  [TrainerType.NORMAN]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["NORMAN"], true, Type.NORMAL)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym"),
-  [TrainerType.WINONA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["WINONA"], false, Type.FLYING)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym"),
-  [TrainerType.TATE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["TATE"], true, Type.PSYCHIC)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym")
-    .setHasDouble("tate_liza_double")
-    .setDoubleTrainerType(TrainerType.LIZA)
-    .setDoubleTitle("gym_leader_double"),
-  [TrainerType.LIZA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["LIZA"], false, Type.PSYCHIC)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym")
-    .setHasDouble("liza_tate_double")
-    .setDoubleTrainerType(TrainerType.TATE)
-    .setDoubleTitle("gym_leader_double"),
-  [TrainerType.JUAN]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["JUAN"], true, Type.WATER)
-    .setBattleBgm("battle_hoenn_gym")
-    .setMixedBattleBgm("battle_hoenn_gym"),
-  [TrainerType.ROARK]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["ROARK"], true, Type.ROCK)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.GARDENIA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["GARDENIA"], false, Type.GRASS)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.MAYLENE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["MAYLENE"], false, Type.FIGHTING)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.CRASHER_WAKE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CRASHER_WAKE"], true, Type.WATER)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.FANTINA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["FANTINA"], false, Type.GHOST)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.BYRON]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BYRON"], true, Type.STEEL)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.CANDICE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CANDICE"], false, Type.ICE)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.VOLKNER]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["VOLKNER"], true, Type.ELECTRIC)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.CILAN]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CILAN"], true, Type.GRASS)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.CHILI]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CHILI"], true, Type.FIRE)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.CRESS]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CRESS"], true, Type.WATER)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.CHEREN]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CHEREN"], true, Type.NORMAL)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.LENORA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["LENORA"], false, Type.NORMAL)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.ROXIE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["ROXIE"], false, Type.POISON)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.BURGH]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BURGH"], true, Type.BUG)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.ELESA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["ELESA"], false, Type.ELECTRIC)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.CLAY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CLAY"], true, Type.GROUND)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.SKYLA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["SKYLA"], false, Type.FLYING)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.BRYCEN]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BRYCEN"], true, Type.ICE)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.DRAYDEN]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["DRAYDEN"], true, Type.DRAGON)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.MARLON]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["MARLON"], true, Type.WATER)
-    .setMixedBattleBgm("battle_unova_gym"),
-  [TrainerType.VIOLA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["VIOLA"], false, Type.BUG)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.GRANT]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["GRANT"], true, Type.ROCK)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.KORRINA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["KORRINA"], false, Type.FIGHTING)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.RAMOS]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["RAMOS"], true, Type.GRASS)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.CLEMONT]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["CLEMONT"], true, Type.ELECTRIC)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.VALERIE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["VALERIE"], false, Type.FAIRY)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.OLYMPIA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["OLYMPIA"], false, Type.PSYCHIC)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.WULFRIC]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["WULFRIC"], true, Type.ICE)
-    .setMixedBattleBgm("battle_kalos_gym"),
-  [TrainerType.MILO]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["MILO"], true, Type.GRASS)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.NESSA]: new TrainerConfig(++t)
-    .setName("Nessa")
-    .initForGymLeader(signatureSpecies["NESSA"], false, Type.WATER)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.KABU]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["KABU"], true, Type.FIRE)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.BEA]: new TrainerConfig(++t)
-    .setName("Bea")
-    .initForGymLeader(signatureSpecies["BEA"], false, Type.FIGHTING)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.ALLISTER]: new TrainerConfig(++t)
-    .setName("Allister")
-    .initForGymLeader(signatureSpecies["ALLISTER"], true, Type.GHOST)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.OPAL]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["OPAL"], false, Type.FAIRY)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.BEDE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BEDE"], true, Type.FAIRY)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.GORDIE]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["GORDIE"], true, Type.ROCK)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.MELONY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["MELONY"], false, Type.ICE)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.PIERS]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["PIERS"], true, Type.DARK)
-    .setHasDouble("piers_marnie_double")
-    .setDoubleTrainerType(TrainerType.MARNIE)
-    .setDoubleTitle("gym_leader_double")
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.MARNIE]: new TrainerConfig(++t)
-    .setName("Marnie")
-    .initForGymLeader(signatureSpecies["MARNIE"], false, Type.DARK)
-    .setHasDouble("marnie_piers_double")
-    .setDoubleTrainerType(TrainerType.PIERS)
-    .setDoubleTitle("gym_leader_double")
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.RAIHAN]: new TrainerConfig(++t)
-    .setName("Raihan")
-    .initForGymLeader(signatureSpecies["RAIHAN"], true, Type.DRAGON)
-    .setMixedBattleBgm("battle_galar_gym"),
-  [TrainerType.KATY]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["KATY"], false, Type.BUG)
-    .setMixedBattleBgm("battle_paldea_gym"),
-  [TrainerType.BRASSIUS]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["BRASSIUS"], true, Type.GRASS)
-    .setMixedBattleBgm("battle_paldea_gym"),
-  [TrainerType.IONO]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["IONO"], false, Type.ELECTRIC)
-    .setMixedBattleBgm("battle_paldea_gym"),
-  [TrainerType.KOFU]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["KOFU"], true, Type.WATER)
-    .setMixedBattleBgm("battle_paldea_gym"),
-  [TrainerType.LARRY]: new TrainerConfig(++t)
-    .setName("Larry")
-    .initForGymLeader(signatureSpecies["LARRY"], true, Type.NORMAL)
-    .setMixedBattleBgm("battle_paldea_gym"),
-  [TrainerType.RYME]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["RYME"], false, Type.GHOST)
-    .setMixedBattleBgm("battle_paldea_gym"),
-  [TrainerType.TULIP]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["TULIP"], false, Type.PSYCHIC)
-    .setMixedBattleBgm("battle_paldea_gym"),
-  [TrainerType.GRUSHA]: new TrainerConfig(++t)
-    .initForGymLeader(signatureSpecies["GRUSHA"], true, Type.ICE)
-    .setMixedBattleBgm("battle_paldea_gym"),
-
-  [TrainerType.LORELEI]: new TrainerConfig((t = TrainerType.LORELEI))
-    .initForEliteFour(signatureSpecies["LORELEI"], false, Type.ICE)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.BRUNO]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["BRUNO"], true, Type.FIGHTING)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.AGATHA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["AGATHA"], false, Type.GHOST)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.LANCE]: new TrainerConfig(++t)
-    .setName("Lance")
-    .initForEliteFour(signatureSpecies["LANCE"], true, Type.DRAGON)
-    .setBattleBgm("battle_kanto_gym")
-    .setMixedBattleBgm("battle_kanto_gym"),
-  [TrainerType.WILL]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["WILL"], true, Type.PSYCHIC)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.KOGA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["KOGA"], true, Type.POISON)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.KAREN]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["KAREN"], false, Type.DARK)
-    .setBattleBgm("battle_johto_gym")
-    .setMixedBattleBgm("battle_johto_gym"),
-  [TrainerType.SIDNEY]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["SIDNEY"], true, Type.DARK)
-    .setMixedBattleBgm("battle_hoenn_elite"),
-  [TrainerType.PHOEBE]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["PHOEBE"], false, Type.GHOST)
-    .setMixedBattleBgm("battle_hoenn_elite"),
-  [TrainerType.GLACIA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["GLACIA"], false, Type.ICE)
-    .setMixedBattleBgm("battle_hoenn_elite"),
-  [TrainerType.DRAKE]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["DRAKE"], true, Type.DRAGON)
-    .setMixedBattleBgm("battle_hoenn_elite"),
-  [TrainerType.AARON]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["AARON"], true, Type.BUG)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.BERTHA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["BERTHA"], false, Type.GROUND)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.FLINT]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["FLINT"], true, Type.FIRE)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.LUCIAN]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["LUCIAN"], true, Type.PSYCHIC)
-    .setBattleBgm("battle_sinnoh_gym")
-    .setMixedBattleBgm("battle_sinnoh_gym"),
-  [TrainerType.SHAUNTAL]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["SHAUNTAL"], false, Type.GHOST)
-    .setMixedBattleBgm("battle_unova_elite"),
-  [TrainerType.MARSHAL]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["MARSHAL"], true, Type.FIGHTING)
-    .setMixedBattleBgm("battle_unova_elite"),
-  [TrainerType.GRIMSLEY]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["GRIMSLEY"], true, Type.DARK)
-    .setMixedBattleBgm("battle_unova_elite"),
-  [TrainerType.CAITLIN]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["CAITLIN"], false, Type.PSYCHIC)
-    .setMixedBattleBgm("battle_unova_elite"),
-  [TrainerType.MALVA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["MALVA"], false, Type.FIRE)
-    .setMixedBattleBgm("battle_kalos_elite"),
-  [TrainerType.SIEBOLD]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["SIEBOLD"], true, Type.WATER)
-    .setMixedBattleBgm("battle_kalos_elite"),
-  [TrainerType.WIKSTROM]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["WIKSTROM"], true, Type.STEEL)
-    .setMixedBattleBgm("battle_kalos_elite"),
-  [TrainerType.DRASNA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["DRASNA"], false, Type.DRAGON)
-    .setMixedBattleBgm("battle_kalos_elite"),
-  [TrainerType.HALA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["HALA"], true, Type.FIGHTING)
-    .setMixedBattleBgm("battle_alola_elite"),
-  [TrainerType.MOLAYNE]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["MOLAYNE"], true, Type.STEEL)
-    .setMixedBattleBgm("battle_alola_elite"),
-  [TrainerType.OLIVIA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["OLIVIA"], false, Type.ROCK)
-    .setMixedBattleBgm("battle_alola_elite"),
-  [TrainerType.ACEROLA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["ACEROLA"], false, Type.GHOST)
-    .setMixedBattleBgm("battle_alola_elite"),
-  [TrainerType.KAHILI]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["KAHILI"], false, Type.FLYING)
-    .setMixedBattleBgm("battle_alola_elite"),
-  [TrainerType.MARNIE_ELITE]: new TrainerConfig(++t)
-    .setName("Marnie")
-    .initForEliteFour(signatureSpecies["MARNIE_ELITE"], false, Type.DARK)
-    .setMixedBattleBgm("battle_galar_elite"),
-  [TrainerType.NESSA_ELITE]: new TrainerConfig(++t)
-    .setName("Nessa")
-    .initForEliteFour(signatureSpecies["NESSA_ELITE"], false, Type.WATER)
-    .setMixedBattleBgm("battle_galar_elite"),
-  [TrainerType.BEA_ELITE]: new TrainerConfig(++t)
-    .setName("Bea")
-    .initForEliteFour(signatureSpecies["BEA_ELITE"], false, Type.FIGHTING)
-    .setMixedBattleBgm("battle_galar_elite"),
-  [TrainerType.ALLISTER_ELITE]: new TrainerConfig(++t)
-    .setName("Allister")
-    .initForEliteFour(signatureSpecies["ALLISTER_ELITE"], true, Type.GHOST)
-    .setMixedBattleBgm("battle_galar_elite"),
-  [TrainerType.RAIHAN_ELITE]: new TrainerConfig(++t)
-    .setName("Raihan")
-    .initForEliteFour(signatureSpecies["RAIHAN_ELITE"], true, Type.DRAGON)
-    .setMixedBattleBgm("battle_galar_elite"),
-  [TrainerType.RIKA]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["RIKA"], false, Type.GROUND)
-    .setMixedBattleBgm("battle_paldea_elite"),
-  [TrainerType.POPPY]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["POPPY"], false, Type.STEEL)
-    .setMixedBattleBgm("battle_paldea_elite"),
-  [TrainerType.LARRY_ELITE]: new TrainerConfig(++t)
-    .setName("Larry")
-    .initForEliteFour(signatureSpecies["LARRY_ELITE"], true, Type.NORMAL, Type.FLYING)
-    .setMixedBattleBgm("battle_paldea_elite"),
-  [TrainerType.HASSEL]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["HASSEL"], true, Type.DRAGON)
-    .setMixedBattleBgm("battle_paldea_elite"),
-  [TrainerType.CRISPIN]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["CRISPIN"], true, Type.FIRE)
-    .setMixedBattleBgm("battle_bb_elite"),
-  [TrainerType.AMARYS]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["AMARYS"], false, Type.STEEL)
-    .setMixedBattleBgm("battle_bb_elite"),
-  [TrainerType.LACEY]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["LACEY"], false, Type.FAIRY)
-    .setMixedBattleBgm("battle_bb_elite"),
-  [TrainerType.DRAYTON]: new TrainerConfig(++t)
-    .initForEliteFour(signatureSpecies["DRAYTON"], true, Type.DRAGON)
-    .setMixedBattleBgm("battle_bb_elite"),
-
-  [TrainerType.BLUE]: new TrainerConfig((t = TrainerType.BLUE))
-    .initForChampion(signatureSpecies["BLUE"], true)
-    .setBattleBgm("battle_kanto_champion")
-    .setMixedBattleBgm("battle_kanto_champion")
-    .setHasDouble("blue_red_double")
-    .setDoubleTrainerType(TrainerType.RED)
-    .setDoubleTitle("champion_double")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.ALAKAZAM], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.PIDGEOT], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // Mega Pidgeot
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.RED]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["RED"], true)
-    .setBattleBgm("battle_johto_champion")
-    .setMixedBattleBgm("battle_johto_champion")
-    .setHasDouble("red_blue_double")
-    .setDoubleTrainerType(TrainerType.BLUE)
-    .setDoubleTitle("champion_double")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.PIKACHU], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 8; // G-Max Pikachu
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [Species.VENUSAUR, Species.CHARIZARD, Species.BLASTOISE],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => {
-          p.formIndex = 1; // Mega Venusaur, Mega Charizard X, or Mega Blastoise
-          p.generateAndPopulateMoveset();
-          p.generateName();
-        },
-      ),
-    ),
-  [TrainerType.LANCE_CHAMPION]: new TrainerConfig(++t)
-    .setName("Lance")
-    .initForChampion(signatureSpecies["LANCE_CHAMPION"], true)
-    .setBattleBgm("battle_johto_champion")
-    .setMixedBattleBgm("battle_johto_champion")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.AERODACTYL], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.LATIAS, Species.LATIOS], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // Mega Latias or Mega Latios
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.STEVEN]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["STEVEN"], true)
-    .setBattleBgm("battle_hoenn_champion_g5")
-    .setMixedBattleBgm("battle_hoenn_champion_g6")
-    .setHasDouble("steven_wallace_double")
-    .setDoubleTrainerType(TrainerType.WALLACE)
-    .setDoubleTitle("champion_double")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.SKARMORY], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.METAGROSS], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // Mega Metagross
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.WALLACE]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["WALLACE"], true)
-    .setBattleBgm("battle_hoenn_champion_g5")
-    .setMixedBattleBgm("battle_hoenn_champion_g6")
-    .setHasDouble("wallace_steven_double")
-    .setDoubleTrainerType(TrainerType.STEVEN)
-    .setDoubleTitle("champion_double")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.PELIPPER], TrainerSlot.TRAINER, true, (p) => {
-        p.abilityIndex = 1; // Drizzle
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.SWAMPERT], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // Mega Swampert
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.CYNTHIA]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["CYNTHIA"], false)
-    .setBattleBgm("battle_sinnoh_champion")
-    .setMixedBattleBgm("battle_sinnoh_champion")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.SPIRITOMB], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.GARCHOMP], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // Mega Garchomp
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.ALDER]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["ALDER"], true)
-    .setHasDouble("alder_iris_double")
-    .setDoubleTrainerType(TrainerType.IRIS)
-    .setDoubleTitle("champion_double")
-    .setBattleBgm("battle_champion_alder")
-    .setMixedBattleBgm("battle_champion_alder")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.BOUFFALANT, Species.BRAVIARY], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    ),
-  [TrainerType.IRIS]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["IRIS"], false)
-    .setBattleBgm("battle_champion_iris")
-    .setMixedBattleBgm("battle_champion_iris")
-    .setHasDouble("iris_alder_double")
-    .setDoubleTrainerType(TrainerType.ALDER)
-    .setDoubleTitle("champion_double")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.DRUDDIGON], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.LAPRAS], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // G-Max Lapras
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.DIANTHA]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["DIANTHA"], false)
-    .setMixedBattleBgm("battle_kalos_champion")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.GOURGEIST], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.GARDEVOIR], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 1; // Mega Gardevoir
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.HAU]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["HAU"], true)
-    .setMixedBattleBgm("battle_alola_champion")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.ALOLA_RAICHU], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    ),
-  [TrainerType.LEON]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["LEON"], true)
-    .setMixedBattleBgm("battle_galar_champion")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc(
-        [Species.RILLABOOM, Species.CINDERACE, Species.INTELEON],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => {
-          p.generateAndPopulateMoveset();
-        },
-      ),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.CHARIZARD], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 3; // G-Max Charizard
-        p.generateAndPopulateMoveset();
-        p.generateName();
-      }),
-    ),
-  [TrainerType.GEETA]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["GEETA"], false)
-    .setMixedBattleBgm("battle_champion_geeta")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.GLIMMORA], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    ),
-  [TrainerType.NEMONA]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["NEMONA"], false)
-    .setMixedBattleBgm("battle_champion_nemona")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.LYCANROC], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = 0; // Midday form
-        p.generateAndPopulateMoveset();
-      }),
-    ),
-  [TrainerType.KIERAN]: new TrainerConfig(++t)
-    .initForChampion(signatureSpecies["KIERAN"], true)
-    .setMixedBattleBgm("battle_champion_kieran")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.POLIWRATH, Species.POLITOED], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-      }),
-    ),
-
-  [TrainerType.RIVAL]: new TrainerConfig((t = TrainerType.RIVAL))
-    .setName("Finn")
-    .setHasGenders("Ivy")
-    .setHasCharSprite()
-    .setTitle("Rival")
-    .setStaticParty()
-    .setEncounterBgm(TrainerType.RIVAL)
-    .setBattleBgm("battle_rival")
-    .setMixedBattleBgm("battle_rival")
-    .setPartyTemplates(trainerPartyTemplates.RIVAL)
-    .setModifierRewardFuncs(
-      () => modifierTypes.SUPER_EXP_CHARM,
-      () => modifierTypes.EXP_SHARE,
-    )
-    .setEventModifierRewardFuncs(
-      () => modifierTypes.SHINY_CHARM,
-      () => modifierTypes.ABILITY_CHARM,
-    )
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc(
-        [
-          Species.BULBASAUR,
-          Species.CHARMANDER,
-          Species.SQUIRTLE,
-          Species.CHIKORITA,
-          Species.CYNDAQUIL,
-          Species.TOTODILE,
-          Species.TREECKO,
-          Species.TORCHIC,
-          Species.MUDKIP,
-          Species.TURTWIG,
-          Species.CHIMCHAR,
-          Species.PIPLUP,
-          Species.SNIVY,
-          Species.TEPIG,
-          Species.OSHAWOTT,
-          Species.CHESPIN,
-          Species.FENNEKIN,
-          Species.FROAKIE,
-          Species.ROWLET,
-          Species.LITTEN,
-          Species.POPPLIO,
-          Species.GROOKEY,
-          Species.SCORBUNNY,
-          Species.SOBBLE,
-          Species.SPRIGATITO,
-          Species.FUECOCO,
-          Species.QUAXLY,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => (p.abilityIndex = 0),
-      ),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [
-          Species.PIDGEY,
-          Species.HOOTHOOT,
-          Species.TAILLOW,
-          Species.STARLY,
-          Species.PIDOVE,
-          Species.FLETCHLING,
-          Species.PIKIPEK,
-          Species.ROOKIDEE,
-          Species.WATTREL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-      ),
-    ),
-  [TrainerType.RIVAL_2]: new TrainerConfig(++t)
-    .setName("Finn")
-    .setHasGenders("Ivy")
-    .setHasCharSprite()
-    .setTitle("Rival")
-    .setStaticParty()
-    .setMoneyMultiplier(1.25)
-    .setEncounterBgm(TrainerType.RIVAL)
-    .setBattleBgm("battle_rival")
-    .setMixedBattleBgm("battle_rival")
-    .setPartyTemplates(trainerPartyTemplates.RIVAL_2)
-    .setModifierRewardFuncs(() => modifierTypes.EXP_SHARE)
-    .setEventModifierRewardFuncs(() => modifierTypes.SHINY_CHARM)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc(
-        [
-          Species.IVYSAUR,
-          Species.CHARMELEON,
-          Species.WARTORTLE,
-          Species.BAYLEEF,
-          Species.QUILAVA,
-          Species.CROCONAW,
-          Species.GROVYLE,
-          Species.COMBUSKEN,
-          Species.MARSHTOMP,
-          Species.GROTLE,
-          Species.MONFERNO,
-          Species.PRINPLUP,
-          Species.SERVINE,
-          Species.PIGNITE,
-          Species.DEWOTT,
-          Species.QUILLADIN,
-          Species.BRAIXEN,
-          Species.FROGADIER,
-          Species.DARTRIX,
-          Species.TORRACAT,
-          Species.BRIONNE,
-          Species.THWACKEY,
-          Species.RABOOT,
-          Species.DRIZZILE,
-          Species.FLORAGATO,
-          Species.CROCALOR,
-          Species.QUAXWELL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => (p.abilityIndex = 0),
-      ),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [
-          Species.PIDGEOTTO,
-          Species.HOOTHOOT,
-          Species.TAILLOW,
-          Species.STARAVIA,
-          Species.TRANQUILL,
-          Species.FLETCHINDER,
-          Species.TRUMBEAK,
-          Species.CORVISQUIRE,
-          Species.WATTREL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-      ),
-    )
-    .setPartyMemberFunc(
-      2,
-      getSpeciesFilterRandomPartyMemberFunc(
-        (species: PokemonSpecies) =>
-          !pokemonEvolutions.hasOwnProperty(species.speciesId) &&
-          !pokemonPrevolutions.hasOwnProperty(species.speciesId) &&
-          species.baseTotal >= 450,
-      ),
-    ),
-  [TrainerType.RIVAL_3]: new TrainerConfig(++t)
-    .setName("Finn")
-    .setHasGenders("Ivy")
-    .setHasCharSprite()
-    .setTitle("Rival")
-    .setStaticParty()
-    .setMoneyMultiplier(1.5)
-    .setEncounterBgm(TrainerType.RIVAL)
-    .setBattleBgm("battle_rival")
-    .setMixedBattleBgm("battle_rival")
-    .setPartyTemplates(trainerPartyTemplates.RIVAL_3)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc(
-        [
-          Species.VENUSAUR,
-          Species.CHARIZARD,
-          Species.BLASTOISE,
-          Species.MEGANIUM,
-          Species.TYPHLOSION,
-          Species.FERALIGATR,
-          Species.SCEPTILE,
-          Species.BLAZIKEN,
-          Species.SWAMPERT,
-          Species.TORTERRA,
-          Species.INFERNAPE,
-          Species.EMPOLEON,
-          Species.SERPERIOR,
-          Species.EMBOAR,
-          Species.SAMUROTT,
-          Species.CHESNAUGHT,
-          Species.DELPHOX,
-          Species.GRENINJA,
-          Species.DECIDUEYE,
-          Species.INCINEROAR,
-          Species.PRIMARINA,
-          Species.RILLABOOM,
-          Species.CINDERACE,
-          Species.INTELEON,
-          Species.MEOWSCARADA,
-          Species.SKELEDIRGE,
-          Species.QUAQUAVAL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => (p.abilityIndex = 0),
-      ),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [
-          Species.PIDGEOT,
-          Species.NOCTOWL,
-          Species.SWELLOW,
-          Species.STARAPTOR,
-          Species.UNFEZANT,
-          Species.TALONFLAME,
-          Species.TOUCANNON,
-          Species.CORVIKNIGHT,
-          Species.KILOWATTREL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-      ),
-    )
-    .setPartyMemberFunc(
-      2,
-      getSpeciesFilterRandomPartyMemberFunc(
-        (species: PokemonSpecies) =>
-          !pokemonEvolutions.hasOwnProperty(species.speciesId) &&
-          !pokemonPrevolutions.hasOwnProperty(species.speciesId) &&
-          species.baseTotal >= 450,
-      ),
-    )
-    .setSpeciesFilter((species) => species.baseTotal >= 540),
-  [TrainerType.RIVAL_4]: new TrainerConfig(++t)
-    .setName("Finn")
-    .setHasGenders("Ivy")
-    .setHasCharSprite()
-    .setTitle("Rival")
-    .setBoss()
-    .setStaticParty()
-    .setMoneyMultiplier(1.75)
-    .setEncounterBgm(TrainerType.RIVAL)
-    .setBattleBgm("battle_rival_2")
-    .setMixedBattleBgm("battle_rival_2")
-    .setPartyTemplates(trainerPartyTemplates.RIVAL_4)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc(
-        [
-          Species.VENUSAUR,
-          Species.CHARIZARD,
-          Species.BLASTOISE,
-          Species.MEGANIUM,
-          Species.TYPHLOSION,
-          Species.FERALIGATR,
-          Species.SCEPTILE,
-          Species.BLAZIKEN,
-          Species.SWAMPERT,
-          Species.TORTERRA,
-          Species.INFERNAPE,
-          Species.EMPOLEON,
-          Species.SERPERIOR,
-          Species.EMBOAR,
-          Species.SAMUROTT,
-          Species.CHESNAUGHT,
-          Species.DELPHOX,
-          Species.GRENINJA,
-          Species.DECIDUEYE,
-          Species.INCINEROAR,
-          Species.PRIMARINA,
-          Species.RILLABOOM,
-          Species.CINDERACE,
-          Species.INTELEON,
-          Species.MEOWSCARADA,
-          Species.SKELEDIRGE,
-          Species.QUAQUAVAL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => (p.abilityIndex = 0),
-      ),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [
-          Species.PIDGEOT,
-          Species.NOCTOWL,
-          Species.SWELLOW,
-          Species.STARAPTOR,
-          Species.UNFEZANT,
-          Species.TALONFLAME,
-          Species.TOUCANNON,
-          Species.CORVIKNIGHT,
-          Species.KILOWATTREL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-      ),
-    )
-    .setPartyMemberFunc(
-      2,
-      getSpeciesFilterRandomPartyMemberFunc(
-        (species: PokemonSpecies) =>
-          !pokemonEvolutions.hasOwnProperty(species.speciesId) &&
-          !pokemonPrevolutions.hasOwnProperty(species.speciesId) &&
-          species.baseTotal >= 450,
-      ),
-    )
-    .setSpeciesFilter((species) => species.baseTotal >= 540)
-    .setGenModifiersFunc((party) => {
-      const starter = party[0];
-      return [
-        modifierTypes
-          .TERA_SHARD()
-          .generateType([], [starter.species.type1])!
-          .withIdFromFunc(modifierTypes.TERA_SHARD)
-          .newModifier(starter) as PersistentModifier,
-      ]; // TODO: is the bang correct?
-    }),
-  [TrainerType.RIVAL_5]: new TrainerConfig(++t)
-    .setName("Finn")
-    .setHasGenders("Ivy")
-    .setHasCharSprite()
-    .setTitle("Rival")
-    .setBoss()
-    .setStaticParty()
-    .setMoneyMultiplier(2.25)
-    .setEncounterBgm(TrainerType.RIVAL)
-    .setBattleBgm("battle_rival_3")
-    .setMixedBattleBgm("battle_rival_3")
-    .setPartyTemplates(trainerPartyTemplates.RIVAL_5)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc(
-        [
-          Species.VENUSAUR,
-          Species.CHARIZARD,
-          Species.BLASTOISE,
-          Species.MEGANIUM,
-          Species.TYPHLOSION,
-          Species.FERALIGATR,
-          Species.SCEPTILE,
-          Species.BLAZIKEN,
-          Species.SWAMPERT,
-          Species.TORTERRA,
-          Species.INFERNAPE,
-          Species.EMPOLEON,
-          Species.SERPERIOR,
-          Species.EMBOAR,
-          Species.SAMUROTT,
-          Species.CHESNAUGHT,
-          Species.DELPHOX,
-          Species.GRENINJA,
-          Species.DECIDUEYE,
-          Species.INCINEROAR,
-          Species.PRIMARINA,
-          Species.RILLABOOM,
-          Species.CINDERACE,
-          Species.INTELEON,
-          Species.MEOWSCARADA,
-          Species.SKELEDIRGE,
-          Species.QUAQUAVAL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => {
-          p.setBoss(true, 2);
-          p.abilityIndex = 0;
-        },
-      ),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [
-          Species.PIDGEOT,
-          Species.NOCTOWL,
-          Species.SWELLOW,
-          Species.STARAPTOR,
-          Species.UNFEZANT,
-          Species.TALONFLAME,
-          Species.TOUCANNON,
-          Species.CORVIKNIGHT,
-          Species.KILOWATTREL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-      ),
-    )
-    .setPartyMemberFunc(
-      2,
-      getSpeciesFilterRandomPartyMemberFunc(
-        (species: PokemonSpecies) =>
-          !pokemonEvolutions.hasOwnProperty(species.speciesId) &&
-          !pokemonPrevolutions.hasOwnProperty(species.speciesId) &&
-          species.baseTotal >= 450,
-      ),
-    )
-    .setSpeciesFilter((species) => species.baseTotal >= 540)
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.RAYQUAZA], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 3);
-        p.pokeball = PokeballType.MASTER_BALL;
-        p.shiny = true;
-        p.variant = 1;
-      }),
-    )
-    .setGenModifiersFunc((party) => {
-      const starter = party[0];
-      return [
-        modifierTypes
-          .TERA_SHARD()
-          .generateType([], [starter.species.type1])!
-          .withIdFromFunc(modifierTypes.TERA_SHARD)
-          .newModifier(starter) as PersistentModifier,
-      ]; //TODO: is the bang correct?
-    }),
-  [TrainerType.RIVAL_6]: new TrainerConfig(++t)
-    .setName("Finn")
-    .setHasGenders("Ivy")
-    .setHasCharSprite()
-    .setTitle("Rival")
-    .setBoss()
-    .setStaticParty()
-    .setMoneyMultiplier(3)
-    .setEncounterBgm("final")
-    .setBattleBgm("battle_rival_3")
-    .setMixedBattleBgm("battle_rival_3")
-    .setPartyTemplates(trainerPartyTemplates.RIVAL_6)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc(
-        [
-          Species.VENUSAUR,
-          Species.CHARIZARD,
-          Species.BLASTOISE,
-          Species.MEGANIUM,
-          Species.TYPHLOSION,
-          Species.FERALIGATR,
-          Species.SCEPTILE,
-          Species.BLAZIKEN,
-          Species.SWAMPERT,
-          Species.TORTERRA,
-          Species.INFERNAPE,
-          Species.EMPOLEON,
-          Species.SERPERIOR,
-          Species.EMBOAR,
-          Species.SAMUROTT,
-          Species.CHESNAUGHT,
-          Species.DELPHOX,
-          Species.GRENINJA,
-          Species.DECIDUEYE,
-          Species.INCINEROAR,
-          Species.PRIMARINA,
-          Species.RILLABOOM,
-          Species.CINDERACE,
-          Species.INTELEON,
-          Species.MEOWSCARADA,
-          Species.SKELEDIRGE,
-          Species.QUAQUAVAL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => {
-          p.setBoss(true, 3);
-          p.abilityIndex = 0;
-          p.generateAndPopulateMoveset();
-        },
-      ),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc(
-        [
-          Species.PIDGEOT,
-          Species.NOCTOWL,
-          Species.SWELLOW,
-          Species.STARAPTOR,
-          Species.UNFEZANT,
-          Species.TALONFLAME,
-          Species.TOUCANNON,
-          Species.CORVIKNIGHT,
-          Species.KILOWATTREL,
-        ],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => {
-          p.setBoss(true, 2);
-          p.generateAndPopulateMoveset();
-        },
-      ),
-    )
-    .setPartyMemberFunc(
-      2,
-      getSpeciesFilterRandomPartyMemberFunc(
-        (species: PokemonSpecies) =>
-          !pokemonEvolutions.hasOwnProperty(species.speciesId) &&
-          !pokemonPrevolutions.hasOwnProperty(species.speciesId) &&
-          species.baseTotal >= 450,
-      ),
-    )
-    .setSpeciesFilter((species) => species.baseTotal >= 540)
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.RAYQUAZA], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss();
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-        p.shiny = true;
-        p.variant = 1;
-        p.formIndex = 1; // Mega Rayquaza
-        p.generateName();
-      }),
-    )
-    .setGenModifiersFunc((party) => {
-      const starter = party[0];
-      return [
-        modifierTypes
-          .TERA_SHARD()
-          .generateType([], [starter.species.type1])!
-          .withIdFromFunc(modifierTypes.TERA_SHARD)
-          .newModifier(starter) as PersistentModifier,
-      ]; // TODO: is the bang correct?
-    }),
-
-  [TrainerType.ROCKET_BOSS_GIOVANNI_1]: new TrainerConfig((t = TrainerType.ROCKET_BOSS_GIOVANNI_1))
-    .setName("Giovanni")
-    .initForEvilTeamLeader("Rocket Boss", [])
-    .setMixedBattleBgm("battle_rocket_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.PERSIAN, Species.ALOLA_PERSIAN], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.MALE;
-      }),
-    )
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.DUGTRIO, Species.ALOLA_DUGTRIO]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.HONCHKROW]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.NIDOKING, Species.NIDOQUEEN]))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.RHYPERIOR]))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.KANGASKHAN], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Kangaskhan
-        p.generateName();
-      }),
-    ),
-  [TrainerType.ROCKET_BOSS_GIOVANNI_2]: new TrainerConfig(++t)
-    .setName("Giovanni")
-    .initForEvilTeamLeader("Rocket Boss", [], true)
-    .setMixedBattleBgm("battle_rocket_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.TYRANITAR, Species.IRON_THORNS], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.HIPPOWDON]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.EXCADRILL, Species.GARCHOMP]))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.KANGASKHAN], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Kangaskhan
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.GASTRODON, Species.SEISMITOAD], TrainerSlot.TRAINER, true, (p) => {
-        //Storm Drain Gastrodon, Water Absorb Seismitoad
-        if (p.species.speciesId === Species.GASTRODON) {
-          p.abilityIndex = 0;
-        } else if (p.species.speciesId === Species.SEISMITOAD) {
-          p.abilityIndex = 2;
-        }
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.MEWTWO], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.MAXIE]: new TrainerConfig(++t)
-    .setName("Maxie")
-    .initForEvilTeamLeader("Magma Boss", [])
-    .setMixedBattleBgm("battle_aqua_magma_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(0, getRandomPartyMemberFunc([Species.MIGHTYENA]))
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.CROBAT, Species.GLISCOR]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.WEEZING, Species.GALAR_WEEZING]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.MAGMORTAR, Species.TORKOAL]))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.FLYGON]))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.CAMERUPT], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Camerupt
-        p.generateName();
-        p.gender = Gender.MALE;
-      }),
-    ),
-  [TrainerType.MAXIE_2]: new TrainerConfig(++t)
-    .setName("Maxie")
-    .initForEvilTeamLeader("Magma Boss", [], true)
-    .setMixedBattleBgm("battle_aqua_magma_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.SOLROCK, Species.TYPHLOSION], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.TORKOAL, Species.NINETALES], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 2; // Drought
-      }),
-    )
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.SHIFTRY, Species.SCOVILLAIN], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 0; // Chlorophyll
-      }),
-    )
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.GREAT_TUSK]))
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.CAMERUPT], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Camerupt
-        p.generateName();
-        p.gender = Gender.MALE;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.GROUDON], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.ARCHIE]: new TrainerConfig(++t)
-    .setName("Archie")
-    .initForEvilTeamLeader("Aqua Boss", [])
-    .setMixedBattleBgm("battle_aqua_magma_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(0, getRandomPartyMemberFunc([Species.LINOONE]))
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.CROBAT, Species.PELIPPER]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.MUK, Species.ALOLA_MUK]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.TENTACRUEL]))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.RELICANTH, Species.WAILORD]))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.SHARPEDO], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Sharpedo
-        p.generateName();
-        p.gender = Gender.MALE;
-      }),
-    ),
-  [TrainerType.ARCHIE_2]: new TrainerConfig(++t)
-    .setName("Archie")
-    .initForEvilTeamLeader("Aqua Boss", [], true)
-    .setMixedBattleBgm("battle_aqua_magma_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.EMPOLEON, Species.LUDICOLO], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.POLITOED, Species.PELIPPER], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 2; // Drizzle
-      }),
-    )
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.BEARTIC, Species.ARMALDO], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 2; // Swift Swim
-      }),
-    )
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.OVERQWIL], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 1; // Swift Swim
-      }),
-    )
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.SHARPEDO], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Sharpedo
-        p.generateName();
-        p.gender = Gender.MALE;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.KYOGRE], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.CYRUS]: new TrainerConfig(++t)
-    .setName("Cyrus")
-    .initForEvilTeamLeader("Galactic Boss", [])
-    .setMixedBattleBgm("battle_galactic_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(0, getRandomPartyMemberFunc([Species.GYARADOS]))
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.HONCHKROW, Species.HISUI_BRAVIARY]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.CROBAT, Species.GLISCOR]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.AZELF, Species.UXIE, Species.MESPRIT]))
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.HOUNDOOM], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Houndoom
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.WEAVILE], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.gender = Gender.MALE;
-      }),
-    ),
-  [TrainerType.CYRUS_2]: new TrainerConfig(++t)
-    .setName("Cyrus")
-    .initForEvilTeamLeader("Galactic Boss", [], true)
-    .setMixedBattleBgm("battle_galactic_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.AZELF, Species.UXIE, Species.MESPRIT], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.ELECTRODE, Species.HISUI_ELECTRODE]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.SALAMENCE, Species.ROARING_MOON]))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.HOUNDOOM], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Houndoom
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.WEAVILE, Species.SNEASLER], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.gender = Gender.MALE;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.DARKRAI], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.GHETSIS]: new TrainerConfig(++t)
-    .setName("Ghetsis")
-    .initForEvilTeamLeader("Plasma Boss", [])
-    .setMixedBattleBgm("battle_plasma_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(0, getRandomPartyMemberFunc([Species.COFAGRIGUS, Species.RUNERIGUS]))
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.BOUFFALANT]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.SEISMITOAD, Species.CARRACOSTA]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.EELEKTROSS, Species.GALVANTULA]))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.VOLCARONA]))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.HYDREIGON], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.gender = Gender.MALE;
-      }),
-    ),
-  [TrainerType.GHETSIS_2]: new TrainerConfig(++t)
-    .setName("Ghetsis")
-    .initForEvilTeamLeader("Plasma Boss", [], true)
-    .setMixedBattleBgm("battle_plasma_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.GENESECT], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = Utils.randSeedInt(4, 1); // Shock, Burn, Chill, or Douse Drive
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.BASCULEGION, Species.JELLICENT], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.MALE;
-        p.formIndex = 0;
-      }),
-    )
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.KINGAMBIT]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.VOLCARONA, Species.SLITHER_WING]))
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.HYDREIGON, Species.IRON_JUGULIS], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        if (p.species.speciesId === Species.HYDREIGON) {
-          p.gender = Gender.MALE;
-        } else if (p.species.speciesId === Species.IRON_JUGULIS) {
-          p.gender = Gender.GENDERLESS;
-        }
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.KYUREM], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.LYSANDRE]: new TrainerConfig(++t)
-    .setName("Lysandre")
-    .initForEvilTeamLeader("Flare Boss", [])
-    .setMixedBattleBgm("battle_flare_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(0, getRandomPartyMemberFunc([Species.MIENSHAO]))
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.HONCHKROW, Species.TALONFLAME]))
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.PYROAR], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.MALE;
-      }),
-    )
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.CLAWITZER, Species.DRAGALGE]))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.HELIOLISK, Species.MALAMAR]))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.GYARADOS], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Gyarados
-        p.generateName();
-        p.gender = Gender.MALE;
-      }),
-    ),
-  [TrainerType.LYSANDRE_2]: new TrainerConfig(++t)
-    .setName("Lysandre")
-    .initForEvilTeamLeader("Flare Boss", [], true)
-    .setMixedBattleBgm("battle_flare_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.SCREAM_TAIL, Species.FLUTTER_MANE], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.PYROAR], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.MALE;
-      }),
-    )
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.IRON_MOTH]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.GOODRA, Species.HISUI_GOODRA]))
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.GYARADOS], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.formIndex = 1; // Mega Gyardos
-        p.generateName();
-        p.gender = Gender.MALE;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.YVELTAL], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.LUSAMINE]: new TrainerConfig(++t)
-    .setName("Lusamine")
-    .initForEvilTeamLeader("Aether Boss", [])
-    .setMixedBattleBgm("battle_aether_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.CLEFABLE], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.FEMALE;
-      }),
-    )
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.LILLIGANT, Species.HISUI_LILLIGANT]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.MILOTIC, Species.PRIMARINA]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.GALAR_SLOWBRO, Species.GALAR_SLOWKING]))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.BEWEAR]))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.NIHILEGO], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ROGUE_BALL;
-      }),
-    ),
-  [TrainerType.LUSAMINE_2]: new TrainerConfig(++t)
-    .setName("Lusamine")
-    .initForEvilTeamLeader("Aether Boss", [], true)
-    .setMixedBattleBgm("battle_aether_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.NIHILEGO], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ROGUE_BALL;
-      }),
-    )
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.MILOTIC, Species.PRIMARINA]))
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.CLEFABLE], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.FEMALE;
-      }),
-    )
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc(
-        [Species.STAKATAKA, Species.CELESTEELA, Species.GUZZLORD],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => {
-          p.generateAndPopulateMoveset();
-          p.pokeball = PokeballType.ROGUE_BALL;
-        },
-      ),
-    )
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.PHEROMOSA], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ROGUE_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.NECROZMA], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.GUZMA]: new TrainerConfig(++t)
-    .setName("Guzma")
-    .initForEvilTeamLeader("Skull Boss", [])
-    .setMixedBattleBgm("battle_skull_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.LOKIX, Species.YANMEGA], TrainerSlot.TRAINER, true, (p) => {
-        //Tinted Lens Lokix, Tinted Lens Yanmega
-        if (p.species.speciesId === Species.LOKIX) {
-          p.abilityIndex = 2;
-        } else if (p.species.speciesId === Species.YANMEGA) {
-          p.abilityIndex = 1;
-        }
-      }),
-    )
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.HERACROSS]))
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.SCIZOR, Species.KLEAVOR], TrainerSlot.TRAINER, true, (p) => {
-        //Technician Scizor, Sharpness Kleavor
-        if (p.species.speciesId === Species.SCIZOR) {
-          p.abilityIndex = 1;
-        } else if (p.species.speciesId === Species.KLEAVOR) {
-          p.abilityIndex = 2;
-        }
-      }),
-    )
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.GALVANTULA, Species.VIKAVOLT]))
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.PINSIR], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1; // Mega Pinsir
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.GOLISOPOD], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.MALE;
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    ),
-  [TrainerType.GUZMA_2]: new TrainerConfig(++t)
-    .setName("Guzma")
-    .initForEvilTeamLeader("Skull Boss", [], true)
-    .setMixedBattleBgm("battle_skull_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.GOLISOPOD], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 2; //Anticipation
-        p.gender = Gender.MALE;
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.SCIZOR, Species.KLEAVOR], TrainerSlot.TRAINER, true, (p) => {
-        //Technician Scizor, Sharpness Kleavor
-        if (p.species.speciesId === Species.SCIZOR) {
-          p.abilityIndex = 1;
-        } else if (p.species.speciesId === Species.KLEAVOR) {
-          p.abilityIndex = 2;
-        }
-      }),
-    )
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.HISUI_SAMUROTT, Species.CRAWDAUNT], TrainerSlot.TRAINER, true, (p) => {
-        p.abilityIndex = 2; //Sharpness Hisui Samurott, Adaptability Crawdaunt
-      }),
-    )
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.BUZZWOLE], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ROGUE_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.XURKITREE], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ROGUE_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.PINSIR], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.formIndex = 1;
-        p.generateAndPopulateMoveset();
-        p.generateName();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    ),
-  [TrainerType.ROSE]: new TrainerConfig(++t)
-    .setName("Rose")
-    .initForEvilTeamLeader("Macro Boss", [])
-    .setMixedBattleBgm("battle_macro_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(0, getRandomPartyMemberFunc([Species.ARCHALUDON]))
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.FERROTHORN, Species.ESCAVALIER]))
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.SIRFETCHD, Species.MR_RIME]))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.CORVIKNIGHT]))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.PERRSERKER, Species.KLINKLANG]))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.COPPERAJAH], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1; // G-Max Copperajah
-        p.generateName();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.gender = Gender.FEMALE;
-      }),
-    ),
-  [TrainerType.ROSE_2]: new TrainerConfig(++t)
-    .setName("Rose")
-    .initForEvilTeamLeader("Macro Boss", [], true)
-    .setMixedBattleBgm("battle_macro_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.ARCHALUDON], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(1, getRandomPartyMemberFunc([Species.AEGISLASH, Species.GHOLDENGO]))
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.DRACOVISH, Species.DRACOZOLT], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 1; //Strong Jaw Dracovish, Hustle Dracozolt
-      }),
-    )
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.MELMETAL]))
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc(
-        [Species.GALAR_ARTICUNO, Species.GALAR_ZAPDOS, Species.GALAR_MOLTRES],
-        TrainerSlot.TRAINER,
-        true,
-        (p) => {
-          p.setBoss(true, 2);
-          p.generateAndPopulateMoveset();
-          p.pokeball = PokeballType.ULTRA_BALL;
-        },
-      ),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.COPPERAJAH], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1; // G-Max Copperajah
-        p.generateName();
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.gender = Gender.FEMALE;
-      }),
-    ),
-  [TrainerType.PENNY]: new TrainerConfig(++t)
-    .setName("Cassiopeia")
-    .initForEvilTeamLeader("Star Boss", [])
-    .setMixedBattleBgm("battle_star_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(0, getRandomPartyMemberFunc([Species.VAPOREON, Species.JOLTEON, Species.FLAREON]))
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.ESPEON, Species.UMBREON], TrainerSlot.TRAINER, true, (p) => {
-        p.abilityIndex = 2; // Magic Bounce Espeon, Inner Focus Umbreon
-        p.generateAndPopulateMoveset();
-      }),
-    )
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.LEAFEON, Species.GLACEON]))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.ROTOM], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.formIndex = Utils.randSeedInt(5, 1); // Heat, Wash, Frost, Fan, or Mow
-      }),
-    )
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.SYLVEON], TrainerSlot.TRAINER, true, (p) => {
-        p.abilityIndex = 2; // Pixilate
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.FEMALE;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.EEVEE], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.formIndex = 2; // G-Max Eevee
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.generateName();
-      }),
-    )
-    .setGenModifiersFunc((party) => {
-      const teraPokemon = party[4];
-      return [
-        modifierTypes
-          .TERA_SHARD()
-          .generateType([], [teraPokemon.species.type1])!
-          .withIdFromFunc(modifierTypes.TERA_SHARD)
-          .newModifier(teraPokemon) as PersistentModifier,
-      ]; //TODO: is the bang correct?
-    }),
-  [TrainerType.PENNY_2]: new TrainerConfig(++t)
-    .setName("Cassiopeia")
-    .initForEvilTeamLeader("Star Boss", [], true)
-    .setMixedBattleBgm("battle_star_boss")
-    .setVictoryBgm("victory_team_plasma")
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.SYLVEON], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.abilityIndex = 2; // Pixilate
-        p.generateAndPopulateMoveset();
-        p.gender = Gender.FEMALE;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.ENTEI, Species.RAIKOU, Species.SUICUNE], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.WALKING_WAKE, Species.GOUGING_FIRE, Species.RAGING_BOLT]))
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.REVAVROOM], TrainerSlot.TRAINER, true, (p) => {
-        p.formIndex = Utils.randSeedInt(5, 1); //Random Starmobile form
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ROGUE_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      4,
-      getRandomPartyMemberFunc([Species.EEVEE], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.formIndex = 2;
-        p.generateName();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.ZAMAZENTA], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    )
-    .setGenModifiersFunc((party) => {
-      const teraPokemon = party[0];
-      return [
-        modifierTypes
-          .TERA_SHARD()
-          .generateType([], [teraPokemon.species.type1])!
-          .withIdFromFunc(modifierTypes.TERA_SHARD)
-          .newModifier(teraPokemon) as PersistentModifier,
-      ]; //TODO: is the bang correct?
-    }),
-  [TrainerType.BUCK]: new TrainerConfig(++t)
-    .setName("Buck")
-    .initForStatTrainer([], true)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.CLAYDOL], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 3);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.VENUSAUR, Species.COALOSSAL], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.GREAT_BALL;
-        if (p.species.speciesId === Species.VENUSAUR) {
-          p.formIndex = 2; // Gmax
-          p.abilityIndex = 2; // Venusaur gets Chlorophyll
-        } else {
-          p.formIndex = 1; // Gmax
-        }
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.AGGRON], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1; // Mega
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      3,
-      getRandomPartyMemberFunc([Species.TORKOAL], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.abilityIndex = 1; // Drought
-      }),
-    )
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.GREAT_TUSK], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.HEATRAN], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.CHERYL]: new TrainerConfig(++t)
-    .setName("Cheryl")
-    .initForStatTrainer([], false)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.BLISSEY], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 3);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.SNORLAX, Species.LAPRAS], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.GREAT_BALL;
-        p.formIndex = 1; // Gmax
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.AUDINO], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1; // Mega
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.GOODRA], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.IRON_HANDS], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.CRESSELIA, Species.ENAMORUS], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        if (p.species.speciesId === Species.ENAMORUS) {
-          p.formIndex = 1; // Therian
-          p.generateName();
-        }
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.MARLEY]: new TrainerConfig(++t)
-    .setName("Marley")
-    .initForStatTrainer([], false)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.ARCANINE], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 3);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.ULTRA_BALL;
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.CINDERACE, Species.INTELEON], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.GREAT_BALL;
-        p.formIndex = 1; // Gmax
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      2,
-      getRandomPartyMemberFunc([Species.AERODACTYL], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1; // Mega
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.DRAGAPULT], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.IRON_BUNDLE], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.REGIELEKI], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.MIRA]: new TrainerConfig(++t)
-    .setName("Mira")
-    .initForStatTrainer([], false)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.ALAKAZAM], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1;
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.GENGAR, Species.HATTERENE], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.GREAT_BALL;
-        p.formIndex = p.species.speciesId === Species.GENGAR ? 2 : 1; // Gmax
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.FLUTTER_MANE], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.HYDREIGON], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.MAGNEZONE], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.LATIOS, Species.LATIAS], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.RILEY]: new TrainerConfig(++t)
-    .setName("Riley")
-    .initForStatTrainer([], true)
-    .setPartyMemberFunc(
-      0,
-      getRandomPartyMemberFunc([Species.LUCARIO], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        p.formIndex = 1;
-        p.pokeball = PokeballType.ULTRA_BALL;
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(
-      1,
-      getRandomPartyMemberFunc([Species.RILLABOOM, Species.CENTISKORCH], TrainerSlot.TRAINER, true, (p) => {
-        p.generateAndPopulateMoveset();
-        p.pokeball = PokeballType.GREAT_BALL;
-        p.formIndex = 1; // Gmax
-        p.generateName();
-      }),
-    )
-    .setPartyMemberFunc(2, getRandomPartyMemberFunc([Species.TYRANITAR], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(3, getRandomPartyMemberFunc([Species.ROARING_MOON], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(4, getRandomPartyMemberFunc([Species.URSALUNA], TrainerSlot.TRAINER, true))
-    .setPartyMemberFunc(
-      5,
-      getRandomPartyMemberFunc([Species.REGIGIGAS, Species.LANDORUS], TrainerSlot.TRAINER, true, (p) => {
-        p.setBoss(true, 2);
-        p.generateAndPopulateMoveset();
-        if (p.species.speciesId === Species.LANDORUS) {
-          p.formIndex = 1; // Therian
-          p.generateName();
-        }
-        p.pokeball = PokeballType.MASTER_BALL;
-      }),
-    ),
-  [TrainerType.VICTOR]: new TrainerConfig(++t)
-    .setTitle("The Winstrates")
-    .setLocalizedName("Victor")
-    .setMoneyMultiplier(1) // The Winstrate trainers have total money multiplier of 6
-    .setPartyTemplates(trainerPartyTemplates.ONE_AVG_ONE_STRONG),
-  [TrainerType.VICTORIA]: new TrainerConfig(++t)
-    .setTitle("The Winstrates")
-    .setLocalizedName("Victoria")
-    .setMoneyMultiplier(1)
-    .setPartyTemplates(trainerPartyTemplates.ONE_AVG_ONE_STRONG),
-  [TrainerType.VIVI]: new TrainerConfig(++t)
-    .setTitle("The Winstrates")
-    .setLocalizedName("Vivi")
-    .setMoneyMultiplier(1)
-    .setPartyTemplates(trainerPartyTemplates.TWO_AVG_ONE_STRONG),
-  [TrainerType.VICKY]: new TrainerConfig(++t)
-    .setTitle("The Winstrates")
-    .setLocalizedName("Vicky")
-    .setMoneyMultiplier(1)
-    .setPartyTemplates(trainerPartyTemplates.ONE_AVG),
-  [TrainerType.VITO]: new TrainerConfig(++t)
-    .setTitle("The Winstrates")
-    .setLocalizedName("Vito")
-    .setMoneyMultiplier(2)
-    .setPartyTemplates(
-      new TrainerPartyCompoundTemplate(
-        new TrainerPartyTemplate(3, PartyMemberStrength.AVERAGE),
-        new TrainerPartyTemplate(2, PartyMemberStrength.STRONG),
-      ),
-    ),
-  [TrainerType.BUG_TYPE_SUPERFAN]: new TrainerConfig(++t)
-    .setMoneyMultiplier(2.25)
-    .setEncounterBgm(TrainerType.ACE_TRAINER)
-    .setPartyTemplates(new TrainerPartyTemplate(2, PartyMemberStrength.AVERAGE)),
-  [TrainerType.EXPERT_POKEMON_BREEDER]: new TrainerConfig(++t)
-    .setMoneyMultiplier(3)
-    .setEncounterBgm(TrainerType.ACE_TRAINER)
-    .setLocalizedName("Expert Pokemon Breeder")
-    .setPartyTemplates(new TrainerPartyTemplate(3, PartyMemberStrength.AVERAGE)),
-  [TrainerType.FUTURE_SELF_M]: new TrainerConfig(++t)
-    .setMoneyMultiplier(0)
-    .setEncounterBgm("mystery_encounter_weird_dream")
-    .setBattleBgm("mystery_encounter_weird_dream")
-    .setMixedBattleBgm("mystery_encounter_weird_dream")
-    .setVictoryBgm("mystery_encounter_weird_dream")
-    .setLocalizedName("Future Self M")
-    .setPartyTemplates(new TrainerPartyTemplate(6, PartyMemberStrength.STRONG)),
-  [TrainerType.FUTURE_SELF_F]: new TrainerConfig(++t)
-    .setMoneyMultiplier(0)
-    .setEncounterBgm("mystery_encounter_weird_dream")
-    .setBattleBgm("mystery_encounter_weird_dream")
-    .setMixedBattleBgm("mystery_encounter_weird_dream")
-    .setVictoryBgm("mystery_encounter_weird_dream")
-    .setLocalizedName("Future Self F")
-    .setPartyTemplates(new TrainerPartyTemplate(6, PartyMemberStrength.STRONG)),
+  ...genericTrainerConfigs,
+  ...evilTeamTrainerConfigs,
+  ...gymLeaderTrainerConfigs,
+  ...eliteFourTrainerConfigs,
+  ...chmapionTrainerConfigs,
+  ...rivalTrainerConfigs,
+  ...evilBossTrainerConfigs,
+  ...mysteryEncounterTrainerConfigs,
 };
